Refactor OpenRouter services to integrate new orchestrators and context builders
- Removed ProcedureContextService from OpenRouterServices and replaced it with IntentDetector, RagContextBuilder, ClarificationService, StreamingOrchestrator, and RedactOrchestrator.
- Updated OpenRouterServicesStreamTest to mock and test the new orchestrators and context builders.
- Adjusted tests to ensure proper handling of streaming events and error scenarios.
- Refactored ProcedureContextService tests to utilize the new IntentDetector and RagContextBuilder.
- Enhanced clarity and maintainability of the code by consolidating context detection logic into dedicated services.
