+++ src/main/java/cat/complai/services/openrouter/OpenRouterServices.java
@@ -1,216 +1,79 @@
 package cat.complai.services.openrouter;
 
-import cat.complai.utilities.http.HttpWrapper;
-import cat.complai.dto.http.OpenRouterStreamStartResult;
-import cat.complai.config.CivicVocabularyConfig;
-import cat.complai.helpers.openrouter.CivicVocabularyService;
-import cat.complai.services.openrouter.validation.InputValidationService;
-import cat.complai.services.openrouter.conversation.ConversationManagementService;
-import cat.complai.services.openrouter.ai.AiResponseProcessingService;
-import cat.complai.services.openrouter.procedure.ProcedureContextService;
-import cat.complai.services.openrouter.procedure.ProcedureContextService.ContextRequirements;
-import cat.complai.services.openrouter.procedure.ProcedureContextService.ProcedureContextResult;
-import cat.complai.services.openrouter.procedure.ProcedureContextService.NewsContextResult;
-import cat.complai.services.openrouter.procedure.ProcedureContextService.CityInfoContextResult;
 import cat.complai.dto.openrouter.AskStreamResult;
-import cat.complai.dto.openrouter.OpenRouterResponseDto;
+import cat.complai.dto.openrouter.ComplainantIdentity;
 import cat.complai.dto.openrouter.OpenRouterErrorCode;
+import cat.complai.dto.openrouter.OpenRouterResponseDto;
+import cat.complai.dto.openrouter.OutputFormat;
 import cat.complai.dto.openrouter.Source;
-import cat.complai.dto.openrouter.sse.SseChunkEvent;
-import cat.complai.dto.openrouter.sse.SseSources;
-import cat.complai.dto.openrouter.sse.SseSourcesEvent;
-import cat.complai.dto.openrouter.sse.SseDoneEvent;
-import cat.complai.dto.openrouter.sse.SseErrorEvent;
-import cat.complai.helpers.openrouter.AskStreamErrorMapper;
 import cat.complai.helpers.openrouter.LanguageDetector;
 import cat.complai.helpers.openrouter.RedactPromptBuilder;
-import cat.complai.helpers.openrouter.SseChunkParser;
-import cat.complai.dto.openrouter.ComplainantIdentity;
-import cat.complai.dto.openrouter.OutputFormat;
-import com.fasterxml.jackson.databind.ObjectMapper;
+import cat.complai.services.openrouter.ai.AiResponseProcessingService;
+import cat.complai.services.openrouter.conversation.ConversationManagementService;
+import cat.complai.services.openrouter.validation.InputValidationService;
 import jakarta.annotation.PreDestroy;
-import jakarta.inject.Singleton;
 import jakarta.inject.Inject;
-import org.reactivestreams.Publisher;
-import reactor.core.publisher.Flux;
+import jakarta.inject.Singleton;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.text.Normalizer;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.OptionalInt;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.TimeUnit;
-import java.util.stream.Collectors;
 
 /**
  * Primary implementation of {@link IOpenRouterService}.
  *
- * <p>Orchestrates RAG context retrieval, conversation history management, AI HTTP calls,
- * response caching, and PDF generation. Delegates to focused helper services rather than
- * containing all logic inline.
- */
-/**
- * Primary implementation of {@link IOpenRouterService}.
- *
- * <p>
- * Orchestrates RAG context retrieval, conversation history management, AI HTTP
- * calls,
- * response caching, and PDF generation. Delegates to focused helper services
- * rather than
- * containing all logic inline.
+ * <p>Orchestrates RAG context retrieval, conversation history management, AI
+ * HTTP calls, response caching, and PDF generation. Delegates intent detection
+ * to {@link IntentDetector}, context building to {@link RagContextBuilder},
+ * ambiguity handling to {@link ClarificationService}, streaming to
+ * {@link StreamingOrchestrator}, and redact workflow to
+ * {@link RedactOrchestrator}.
  */
 @Singleton
 public class OpenRouterServices implements IOpenRouterService {
 
     private static final Logger log = LoggerFactory.getLogger(OpenRouterServices.class);
+
     private final InputValidationService validationService;
     private final ConversationManagementService conversationService;
     private final AiResponseProcessingService aiResponseService;
-    private final ProcedureContextService procedureContextService;
+    private final IntentDetector intentDetector;
+    private final RagContextBuilder ragContextBuilder;
+    private final ClarificationService clarificationService;
+    private final StreamingOrchestrator streamingOrchestrator;
+    private final RedactOrchestrator redactOrchestrator;
     private final RedactPromptBuilder promptBuilder;
-    private final HttpWrapper httpWrapper;
-    private final ObjectMapper objectMapper;
     private final ExecutorService ragExecutor;
-    private final CivicVocabularyService civicVocabularyService;
-    private final CivicVocabularyConfig civicVocabularyConfig;
-
-    /**
-     * Estimates token count for text using OpenAI's rule of thumb: ~1 token per 4
-     * characters.
-     */
-    private static int estimateTokenCount(String text) {
-        if (text == null || text.isEmpty()) {
-            return 0;
-        }
-        return Math.max(1, text.length() / 4);
-    }
-
-    /**
-     * Compute a deterministic hash of RAG sources for cache key generation.
-     * 
-     * PRIVACY NOTE: Hashing the sources is safe for caching because:
-     * - Sources contain only public document titles and URLs
-     * - No user query text or conversation history included
-     * - Hash is deterministic: same sources always produce same hash
-     * - Different procedure/event sets produce different hashes
-     * 
-     * @param sources The list of Source objects from RAG
-     * @return A hash of the sources (0 if empty)
-     */
-    private static long computeSourcesHash(List<Source> sources) {
-        if (sources == null || sources.isEmpty()) {
-            return 0;
-        }
-
-        // Create a deterministic hash from source titles and URLs
-        StringBuilder sb = new StringBuilder();
-        for (Source source : sources) {
-            String title = source.getTitle();
-            String url = source.getUrl();
-            if (title != null)
-                sb.append(title).append("|");
-            if (url != null)
-                sb.append(url).append("|");
-        }
-
-        // Use Java's default hashCode (not cryptographically secure, but good for
-        // caching)
-        return sb.toString().hashCode();
-    }
-
-    private static OptionalInt resolveClarificationAnswer(
-            String answer, List<ConversationManagementService.ClarificationCandidate> candidates) {
-        if (answer == null || answer.isBlank() || candidates == null || candidates.isEmpty()) {
-            return OptionalInt.empty();
-        }
-        String normalized = Normalizer.normalize(answer.toLowerCase(Locale.ROOT), Normalizer.Form.NFD)
-                .replaceAll("\\p{InCombiningDiacriticalMarks}+", "");
-
-        String trimmed = normalized.trim();
-        if (trimmed.length() == 1) {
-            char c = trimmed.charAt(0);
-            if (c >= '1' && c <= '9') {
-                int idx = c - '1';
-                if (idx < candidates.size()) {
-                    return OptionalInt.of(idx);
-                }
-            }
-        }
-
-        Map<String, Integer> ordinals = new HashMap<>();
-        ordinals.putAll(Map.of("primer", 0, "primera", 0, "first", 0, "1st", 0, "un", 0,
-                "segon", 1, "segunda", 1, "second", 1, "2nd", 1, "dos", 1));
-        ordinals.putAll(Map.of("tercer", 2, "tercera", 2, "third", 2, "3rd", 2, "tres", 2));
-        for (Map.Entry<String, Integer> entry : ordinals.entrySet()) {
-            if (normalized.contains(entry.getKey()) && entry.getValue() < candidates.size()) {
-                return OptionalInt.of(entry.getValue());
-            }
-        }
-
-        for (int i = 0; i < candidates.size(); i++) {
-            String title = candidates.get(i).title();
-            if (title != null) {
-                String normalizedTitle = Normalizer.normalize(title.toLowerCase(Locale.ROOT), Normalizer.Form.NFD)
-                        .replaceAll("\\p{InCombiningDiacriticalMarks}+", "");
-                if (normalized.contains(normalizedTitle)) {
-                    return OptionalInt.of(i);
-                }
-            }
-        }
-
-        return OptionalInt.empty();
-    }
-
-    /**
-     * Validates and adds sources to the merged list, logging warnings for missing
-     * URLs.
-     * Missing URLs do not prevent the source from being added — logging only.
-     * 
-     * @param mergedSources  The list to add sources to
-     * @param contextSources The sources from a context (procedure, event, news,
-     *                       etc.)
-     * @param contextType    The type of context (for logging): "procedure",
-     *                       "event", "news", "city_info"
-     */
-    private void validateAndAddSources(List<Source> mergedSources, List<Source> contextSources, String contextType) {
-        if (contextSources == null || contextSources.isEmpty()) {
-            return;
-        }
-
-        for (Source source : contextSources) {
-            // Check for missing URL and log warning
-            if (source.getUrl() == null || source.getUrl().isBlank()) {
-                String title = source.getTitle() != null ? source.getTitle() : "<no title>";
-                log.warn("Missing URL for {} item: {}", contextType, title);
-            }
-            // Always add the source, even if URL is missing
-            mergedSources.add(source);
-        }
-    }
 
     @Inject
     public OpenRouterServices(InputValidationService validationService,
-            ConversationManagementService conversationService,
-            AiResponseProcessingService aiResponseService,
-            ProcedureContextService procedureContextService,
-            RedactPromptBuilder promptBuilder,
-            HttpWrapper httpWrapper,
-            ObjectMapper objectMapper,
-            CivicVocabularyService civicVocabularyService,
-            CivicVocabularyConfig civicVocabularyConfig) {
+                              ConversationManagementService conversationService,
+                              AiResponseProcessingService aiResponseService,
+                              IntentDetector intentDetector,
+                              RagContextBuilder ragContextBuilder,
+                              ClarificationService clarificationService,
+                              StreamingOrchestrator streamingOrchestrator,
+                              RedactOrchestrator redactOrchestrator,
+                              RedactPromptBuilder promptBuilder) {
         this.validationService = validationService;
         this.conversationService = conversationService;
         this.aiResponseService = aiResponseService;
-        this.procedureContextService = procedureContextService;
+        this.intentDetector = intentDetector;
+        this.ragContextBuilder = ragContextBuilder;
+        this.clarificationService = clarificationService;
+        this.streamingOrchestrator = streamingOrchestrator;
+        this.redactOrchestrator = redactOrchestrator;
         this.promptBuilder = promptBuilder;
-        this.httpWrapper = httpWrapper;
-        this.objectMapper = objectMapper;
         this.ragExecutor = createRagExecutor();
-        this.civicVocabularyService = civicVocabularyService;
-        this.civicVocabularyConfig = civicVocabularyConfig;
     }
 
     @PreDestroy
@@ -219,13 +82,17 @@ public class OpenRouterServices implements IOpenRouterService {
     }
 
     // -------------------------------------------------------------------------
-    // IOpenRouterService
+    // IOpenRouterService — validateRedactInput
     // -------------------------------------------------------------------------
 
     @Override
     public Optional<OpenRouterResponseDto> validateRedactInput(String complaint) {
-        return validationService.validateRedactInput(complaint);
+        return redactOrchestrator.validateRedactInput(complaint);
     }
+
+    // -------------------------------------------------------------------------
+    // IOpenRouterService — ask (synchronous)
+    // -------------------------------------------------------------------------
 
     @Override
     public OpenRouterResponseDto ask(String question, String conversationId, String cityId) {
@@ -242,85 +109,85 @@ public class OpenRouterServices implements IOpenRouterService {
         String detectedLanguage = RedactPromptBuilder.normalizeLanguageCode(
                 LanguageDetector.detect(question));
 
-        if (procedureContextService.requiresEventDateWindowClarification(question, cityId)) {
+        if (intentDetector.requiresEventDateWindowClarification(question, cityId)) {
             String clarificationMessage = buildEventDateWindowClarificationMessage(detectedLanguage);
             if (conversationId != null && !conversationId.isBlank()) {
                 conversationService.updateConversationHistory(conversationId, question, clarificationMessage);
             }
-            log.info("ask() event date-window clarification triggered - conversationId={} city={}", conversationId,
-                    cityId);
+            log.info("ask() event date-window clarification triggered - conversationId={} city={}", conversationId, cityId);
             return new OpenRouterResponseDto(true, clarificationMessage, null, 200, OpenRouterErrorCode.NONE, null,
                     List.of());
         }
 
         if (conversationId != null && !conversationId.isBlank()) {
             List<ConversationManagementService.ClarificationCandidate> pending =
-                conversationService.getPendingClarification(conversationId);
+                    conversationService.getPendingClarification(conversationId);
             if (pending != null && !pending.isEmpty()) {
                 conversationService.clearPendingClarification(conversationId);
-                OptionalInt resolved = resolveClarificationAnswer(question, pending);
+                OptionalInt resolved = clarificationService.resolveClarificationAnswer(question, pending);
                 if (resolved.isPresent()) {
                     ConversationManagementService.ClarificationCandidate chosen =
-                        pending.get(resolved.getAsInt());
+                            pending.get(resolved.getAsInt());
                     ProcedureContextResult resolvedCtx =
-                        procedureContextService.buildProcedureContextResultForId(
-                            chosen.procedureId(), cityId);
+                            ragContextBuilder.buildProcedureContextResultForId(chosen.procedureId(), cityId);
+
                     List<Map<String, Object>> resolvedMessages = new ArrayList<>();
                     resolvedMessages.add(Map.of("role", "system", "content",
-                        promptBuilder.getSystemMessage(cityId, detectedLanguage)));
+                            promptBuilder.getSystemMessage(cityId, detectedLanguage)));
                     if (resolvedCtx != null && resolvedCtx.getContextBlock() != null) {
                         resolvedMessages.add(Map.of("role", "system", "content", resolvedCtx.getContextBlock()));
                     }
                     var resolvedHistory = conversationService.getConversationHistory(conversationId);
                     conversationService.addToMessages(resolvedMessages, resolvedHistory);
                     resolvedMessages.add(Map.of("role", "user",
-                        "content", question != null ? question.trim() : ""));
+                            "content", question != null ? question.trim() : ""));
+
                     long resolvedProcHash = computeSourcesHash(
-                        resolvedCtx != null ? resolvedCtx.getSources() : List.of());
+                            resolvedCtx != null ? resolvedCtx.getSources() : List.of());
                     OpenRouterResponseDto resolvedResponse =
-                        aiResponseService.callOpenRouterAndExtract(
-                            resolvedMessages, cityId, resolvedProcHash, 0L);
+                            aiResponseService.callOpenRouterAndExtract(
+                                    resolvedMessages, cityId, resolvedProcHash, 0L);
+
                     List<Source> resolvedSources = new ArrayList<>();
-                    validateAndAddSources(resolvedSources,
-                        resolvedCtx != null ? resolvedCtx.getSources() : null, "procedure");
+                    ragContextBuilder.validateAndAddSources(resolvedSources,
+                            resolvedCtx != null ? resolvedCtx.getSources() : null, "procedure");
                     if (resolvedResponse.isSuccess() && !resolvedSources.isEmpty()) {
                         List<Source> finalSources =
-                            procedureContextService.deDuplicateAndOrderSources(resolvedSources);
+                                ragContextBuilder.deDuplicateAndOrderSources(resolvedSources);
                         resolvedResponse = new OpenRouterResponseDto(
-                            resolvedResponse.isSuccess(), resolvedResponse.getMessage(),
-                            resolvedResponse.getError(), resolvedResponse.getStatusCode(),
-                            resolvedResponse.getErrorCode(), resolvedResponse.getPdfData(),
-                            finalSources);
+                                resolvedResponse.isSuccess(), resolvedResponse.getMessage(),
+                                resolvedResponse.getError(), resolvedResponse.getStatusCode(),
+                                resolvedResponse.getErrorCode(), resolvedResponse.getPdfData(),
+                                finalSources);
                     }
                     if (resolvedResponse.getMessage() != null) {
                         conversationService.updateConversationHistory(
-                            conversationId, question, resolvedResponse.getMessage());
+                                conversationId, question, resolvedResponse.getMessage());
                     }
                     log.info("ask() clarification resolved — conversationId={} procedureId={}",
-                        conversationId, chosen.procedureId());
+                            conversationId, chosen.procedureId());
                     return resolvedResponse;
                 }
-                // Resolution failed: fall through to normal RAG (cache already cleared)
                 log.info("ask() clarification unresolvable — falling back to normal RAG conversationId={}",
-                    conversationId);
+                        conversationId);
             }
         }
 
         if (conversationId != null && !conversationId.isBlank()) {
-            Optional<ProcedureContextService.ProcedureAmbiguityResult> ambiguity =
-                procedureContextService.detectProcedureAmbiguity(question, cityId);
+            Optional<ProcedureAmbiguityResult> ambiguity =
+                    clarificationService.detectProcedureAmbiguity(question, cityId);
             if (ambiguity.isPresent()) {
                 List<ConversationManagementService.ClarificationCandidate> candidates =
-                    ambiguity.get().candidates();
+                        ambiguity.get().candidates();
                 conversationService.storePendingClarification(conversationId, candidates);
                 List<String> titles = candidates.stream().map(c -> c.title()).toList();
                 String clarificationMsg = promptBuilder.buildProcedureClarificationMessage(
-                    titles, detectedLanguage, cityId);
+                        titles, detectedLanguage, cityId);
                 conversationService.updateConversationHistory(conversationId, question, clarificationMsg);
                 log.info("ask() procedure ambiguity clarification triggered — conversationId={} candidateCount={}",
-                    conversationId, candidates.size());
+                        conversationId, candidates.size());
                 return new OpenRouterResponseDto(true, clarificationMsg, null, 200,
-                    OpenRouterErrorCode.NONE, null, List.of());
+                        OpenRouterErrorCode.NONE, null, List.of());
             }
         }
 
@@ -328,7 +195,7 @@ public class OpenRouterServices implements IOpenRouterService {
 
         RagContexts ragContexts = buildRagContexts(question, cityId, conversationId, "ask()");
         ProcedureContextResult procCtx = ragContexts.procedureContext();
-        ProcedureContextService.EventContextResult eventCtx = ragContexts.eventContext();
+        EventContextResult eventCtx = ragContexts.eventContext();
         NewsContextResult newsCtx = ragContexts.newsContext();
         CityInfoContextResult cityInfoCtx = ragContexts.cityInfoContext();
 
@@ -342,25 +209,19 @@ public class OpenRouterServices implements IOpenRouterService {
                     List.of());
         }
 
-        // Add procedure context if available
         if (procCtx != null && procCtx.getContextBlock() != null) {
             messages.add(Map.of("role", "system", "content", procCtx.getContextBlock()));
         }
-
-        // Add event context if available
         if (eventCtx != null && eventCtx.getContextBlock() != null) {
             messages.add(Map.of("role", "system", "content", eventCtx.getContextBlock()));
         }
-
         if (newsCtx != null && newsCtx.getContextBlock() != null) {
             messages.add(Map.of("role", "system", "content", newsCtx.getContextBlock()));
         }
-
         if (cityInfoCtx != null && cityInfoCtx.getContextBlock() != null) {
             messages.add(Map.of("role", "system", "content", cityInfoCtx.getContextBlock()));
         }
 
-        // Add conversation history
         var history = conversationService.getConversationHistory(conversationId);
         conversationService.addToMessages(messages, history);
 
@@ -400,10 +261,9 @@ public class OpenRouterServices implements IOpenRouterService {
         log.debug(
                 "ask() CONTEXT METRICS — systemTokens={} historyTokens={} userTokens={} totalTokens={} historyTurns={} conversationId={}",
                 systemTokens, historyTokens, userTokens, totalTokens, historyTurns, conversationId);
-
         log.debug("ask() messages prepared — messageCount={} conversationId={}", messages.size(), conversationId);
 
-        // Calculate context hashes for caching (deterministic, reproducible)
+        // Calculate context hashes for caching
         long procContextHash = computeSourcesHash(procCtx != null ? procCtx.getSources() : List.of());
         List<Source> eventAndNewsSourcesForHash = new ArrayList<>();
         if (eventCtx != null && eventCtx.getSources() != null) {
@@ -417,21 +277,22 @@ public class OpenRouterServices implements IOpenRouterService {
         }
         long eventContextHash = computeSourcesHash(eventAndNewsSourcesForHash);
 
-        OpenRouterResponseDto response = aiResponseService.callOpenRouterAndExtract(messages, cityId, procContextHash,
-                eventContextHash);
+        OpenRouterResponseDto response = aiResponseService.callOpenRouterAndExtract(
+                messages, cityId, procContextHash, eventContextHash);
 
-        // Collect all sources WITHOUT deduplication - symmetric handling
-        // Procedure and event sources are treated equally at collection time
+        // Collect all sources
         List<Source> mergedSources = new ArrayList<>();
-        validateAndAddSources(mergedSources, procCtx != null ? procCtx.getSources() : null, "procedure");
-        validateAndAddSources(mergedSources, eventCtx != null ? eventCtx.getSources() : null, "event");
-        validateAndAddSources(mergedSources, newsCtx != null ? newsCtx.getSources() : null, "news");
-        validateAndAddSources(mergedSources, cityInfoCtx != null ? cityInfoCtx.getSources() : null, "city_info");
+        ragContextBuilder.validateAndAddSources(mergedSources,
+                procCtx != null ? procCtx.getSources() : null, "procedure");
+        ragContextBuilder.validateAndAddSources(mergedSources,
+                eventCtx != null ? eventCtx.getSources() : null, "event");
+        ragContextBuilder.validateAndAddSources(mergedSources,
+                newsCtx != null ? newsCtx.getSources() : null, "news");
+        ragContextBuilder.validateAndAddSources(mergedSources,
+                cityInfoCtx != null ? cityInfoCtx.getSources() : null, "city_info");
 
-        // Deduplicate ONCE after all sources are collected
-        // This ensures symmetric handling and prevents double deduplication
         if (response.isSuccess() && !mergedSources.isEmpty()) {
-            List<Source> finalSources = procedureContextService.deDuplicateAndOrderSources(mergedSources);
+            List<Source> finalSources = ragContextBuilder.deDuplicateAndOrderSources(mergedSources);
             response = new OpenRouterResponseDto(
                     response.isSuccess(),
                     response.getMessage(),
@@ -448,424 +309,32 @@ public class OpenRouterServices implements IOpenRouterService {
         return response;
     }
 
+    // -------------------------------------------------------------------------
+    // IOpenRouterService — redactComplaint (delegates to RedactOrchestrator)
+    // -------------------------------------------------------------------------
+
     @Override
     public OpenRouterResponseDto redactComplaint(String complaint, OutputFormat format, String conversationId,
-            ComplainantIdentity identity, String cityId) {
-        int inputLength = complaint != null ? complaint.length() : 0;
-        boolean identityProvided = identity != null && identity.isPartiallyProvided();
-        log.info("redactComplaint() called — conversationId={} inputLength={} format={} identityProvided={} city={}",
-                conversationId, inputLength, format, identityProvided, cityId);
-
-        var validationError = validationService.validateRedactInput(complaint);
-        if (validationError.isPresent()) {
-            log.debug("redactComplaint() rejected — reason={} conversationId={}", validationError.get().getError(),
-                    conversationId);
-            return validationError.get();
-        }
-
-        // Expand complaint with civic vocabulary for better RAG retrieval
-        String expandedComplaint = complaint;
-        if (civicVocabularyConfig != null && civicVocabularyService != null
-                && civicVocabularyConfig.isEnabled() && complaint != null && !complaint.isBlank()) {
-            String detectedLang = LanguageDetector.detect(complaint);
-            if (!"CA".equalsIgnoreCase(detectedLang)) {
-                String vocabLang = switch (detectedLang.toUpperCase()) {
-                    case "EN" -> "en";
-                    case "ES" -> "es";
-                    case "FR" -> "fr";
-                    default -> null;
-                };
-                if (vocabLang != null) {
-                    expandedComplaint = civicVocabularyService.expandQuery(complaint, vocabLang);
-                    log.debug("redactComplaint() expanded complaint with civic vocabulary — originalLength={} expandedLength={}",
-                            complaint.length(), expandedComplaint.length());
-                }
-            }
-        }
-
-        List<Map<String, Object>> messages = new ArrayList<>();
-        messages.add(Map.of("role", "system", "content", promptBuilder.getSystemMessage(cityId)));
-
-        // Only add procedure context if we have complete identity (ready to draft)
-        // Skip RAG search during identity collection to improve latency
-        String contextBlock;
-        boolean hasCompleteIdentity = identity != null && identity.isComplete();
-        if (hasCompleteIdentity) {
-            contextBlock = promptBuilder.buildProcedureContextBlock(expandedComplaint, cityId);
-            if (contextBlock != null) {
-                messages.add(Map.of("role", "system", "content", contextBlock));
-            }
-        }
-        // Add conversation history
-        var history = conversationService.getConversationHistory(conversationId);
-        conversationService.addToMessages(messages, history);
-
-        // Choose the prompt based on whether we have a complete identity.
-        // When identity is incomplete, the AI is instructed to ask for the missing
-        // fields first.
-        // The caller is expected to collect the AI's question, present it to the user,
-        // and
-        // re-submit with the full identity in a subsequent request.
-        String userPrompt = "";
-        final boolean identityComplete = identity != null && identity.isComplete();
-        if (identityComplete) {
-            // On the second turn the user's message contains the identity data they typed
-            // (e.g.
-            // "My name is Raul Torres, DNI 49872354C, I live at Carrer Major 10"). The
-            // original
-            // complaint was saved on the first turn. We combine both so that any optional
-            // contact
-            // details the user included in their identity reply (address, phone, etc.) are
-            // visible
-            // to the prompt and can be included in the letter.
-            String originalComplaint = conversationService.getPendingComplaint(conversationId);
-            if (originalComplaint != null) {
-                conversationService.clearPendingComplaint(conversationId);
-                log.debug("redactComplaint() resumed stored complaint — conversationId={} originalLength={}",
-                        conversationId, originalComplaint.length());
-            }
-            String complaintForPrompt = (originalComplaint != null)
-                    ? originalComplaint + "\n\n" + complaint
-                    : complaint;
-            if (complaintForPrompt != null) {
-                userPrompt = promptBuilder.buildRedactPromptWithIdentity(complaintForPrompt, identity, cityId);
-            }
-        } else {
-            if (conversationId != null && !conversationId.isBlank()) {
-                conversationService.storePendingComplaint(conversationId, complaint);
-                log.debug("redactComplaint() saved complaint for identity follow-up — conversationId={}",
-                        conversationId);
-            }
-            userPrompt = promptBuilder.buildRedactPromptRequestingIdentity(complaint, identity, cityId);
-        }
-
-        messages.add(Map.of("role", "user", "content", userPrompt));
-
-        log.debug("redactComplaint() messages prepared — messageCount={} identityComplete={} conversationId={}",
-                messages.size(), identityComplete, conversationId);
-
-        // No procedure/event context for complaints, use default cache key
-        OpenRouterResponseDto aiDto = aiResponseService.callOpenRouterAndExtract(messages, cityId, 0, 0);
-
-        // Process the AI response using the dedicated service
-        OpenRouterResponseDto processedResponse = aiResponseService.processComplaintResponse(aiDto, identityComplete);
-
-        // Update conversation history if we have a valid response
-        if (conversationId != null && !conversationId.isBlank() && processedResponse.getMessage() != null) {
-            conversationService.updateConversationHistory(conversationId, userPrompt, processedResponse.getMessage());
-        }
-
-        return processedResponse;
+                                                  ComplainantIdentity identity, String cityId) {
+        return redactOrchestrator.redactComplaint(complaint, format, conversationId, identity, cityId);
     }
+
+    // -------------------------------------------------------------------------
+    // IOpenRouterService — streamAsk (delegates to StreamingOrchestrator)
+    // -------------------------------------------------------------------------
 
     @Override
     public AskStreamResult streamAsk(String question, String conversationId, String cityId) {
-        // 1. Validate input — emit error event if invalid
-        var validationError = validationService.validateQuestion(question);
-        if (validationError.isPresent()) {
-            String errorText = validationError.get().getError();
-            errorText = errorText != null ? errorText : "Invalid input.";
-            try {
-                SseErrorEvent errorEvent = new SseErrorEvent(errorText, OpenRouterErrorCode.VALIDATION.getCode());
-                String json = objectMapper.writeValueAsString(errorEvent);
-                return new AskStreamResult.Success(Flux.just(json));
-            } catch (Exception e) {
-                log.error("streamAsk() failed to serialize validation error", e);
-                return new AskStreamResult.Error(AskStreamErrorMapper.toResponseDto(e));
-            }
-        }
-
-        // 2. Single-language system message + RAG
-        String detectedLanguage = RedactPromptBuilder.normalizeLanguageCode(LanguageDetector.detect(question));
-
-        if (procedureContextService.requiresEventDateWindowClarification(question, cityId)) {
-            String clarificationMessage = buildEventDateWindowClarificationMessage(detectedLanguage);
-            log.info("streamAsk() event date-window clarification triggered - conversationId={} city={}",
-                    conversationId, cityId);
-            return buildFallbackStreamResult(clarificationMessage, conversationId, question);
-        }
-
-        if (conversationId != null && !conversationId.isBlank()) {
-            List<ConversationManagementService.ClarificationCandidate> streamPending =
-                conversationService.getPendingClarification(conversationId);
-            if (streamPending != null && !streamPending.isEmpty()) {
-                conversationService.clearPendingClarification(conversationId);
-                OptionalInt streamResolved = resolveClarificationAnswer(question, streamPending);
-                if (streamResolved.isPresent()) {
-                    ConversationManagementService.ClarificationCandidate streamChosen =
-                        streamPending.get(streamResolved.getAsInt());
-                    ProcedureContextResult streamResolvedCtx =
-                        procedureContextService.buildProcedureContextResultForId(
-                            streamChosen.procedureId(), cityId);
-                    List<Map<String, Object>> streamResolvedMsgs = new ArrayList<>();
-                    streamResolvedMsgs.add(Map.of("role", "system", "content",
-                        promptBuilder.getSystemMessage(cityId, detectedLanguage)));
-                    if (streamResolvedCtx != null && streamResolvedCtx.getContextBlock() != null) {
-                        streamResolvedMsgs.add(Map.of("role", "system", "content",
-                            streamResolvedCtx.getContextBlock()));
-                    }
-                    var streamResolvedHistory = conversationService.getConversationHistory(conversationId);
-                    conversationService.addToMessages(streamResolvedMsgs, streamResolvedHistory);
-                    streamResolvedMsgs.add(Map.of("role", "user",
-                        "content", question != null ? question.trim() : ""));
-
-                    List<Source> streamResolvedSrcs = new ArrayList<>();
-                    validateAndAddSources(streamResolvedSrcs,
-                        streamResolvedCtx != null ? streamResolvedCtx.getSources() : null, "procedure");
-                    List<SseSources> streamResolvedAllSources = streamResolvedSrcs.stream()
-                        .map(s -> new SseSources(s.getTitle(), s.getUrl()))
-                        .collect(Collectors.toList());
-
-                    final String srCapturedQuestion = question;
-                    final String srCapturedConvId = conversationId;
-                    final StringBuilder srAssembled = new StringBuilder();
-                    final java.util.concurrent.atomic.AtomicBoolean srHasEmitted =
-                        new java.util.concurrent.atomic.AtomicBoolean(false);
-
-                    log.info("streamAsk() clarification resolved — conversationId={} procedureId={}",
-                        conversationId, streamChosen.procedureId());
-
-                    OpenRouterStreamStartResult srStreamStart =
-                        httpWrapper.streamFromOpenRouter(streamResolvedMsgs);
-                    if (srStreamStart instanceof OpenRouterStreamStartResult.Error srErr) {
-                        return new AskStreamResult.Error(AskStreamErrorMapper.toResponseDto(srErr.failure()));
-                    }
-                    Publisher<String> srStream =
-                        ((OpenRouterStreamStartResult.Success) srStreamStart).stream();
-                    return new AskStreamResult.Success(Flux.from(srStream)
-                        .handle((raw, sink) -> {
-                            SseChunkParser.ParseResult parsed = SseChunkParser.parseLine(raw);
-                            if (parsed.state() == SseChunkParser.ParseState.MALFORMED) {
-                                sink.error(new IllegalStateException("Malformed upstream stream chunk"));
-                                return;
-                            }
-                            if (parsed.state() == SseChunkParser.ParseState.DELTA && parsed.delta() != null
-                                    && !parsed.delta().isEmpty()) {
-                                sink.next(parsed.delta());
-                            }
-                        })
-                        .cast(String.class)
-                        .doOnNext(delta -> {
-                            srAssembled.append(delta);
-                            srHasEmitted.set(true);
-                        })
-                        .map(delta -> {
-                            try {
-                                SseChunkEvent chunkEvt = new SseChunkEvent(delta);
-                                return objectMapper.writeValueAsString(chunkEvt);
-                            } catch (Exception e) {
-                                log.error("Failed to serialize resolved chunk", e);
-                                throw new RuntimeException("Chunk serialization failed", e);
-                            }
-                        })
-                        .concatWith(produceSourcesAndDone(streamResolvedAllSources, srCapturedConvId))
-                        .doOnComplete(() -> {
-                            String fullText = srAssembled.toString();
-                            if (srCapturedConvId != null && !srCapturedConvId.isBlank() && !fullText.isBlank()) {
-                                conversationService.updateConversationHistory(
-                                    srCapturedConvId, srCapturedQuestion, fullText);
-                            }
-                            log.info("streamAsk() resolved clarification completed — conversationId={}",
-                                srCapturedConvId);
-                        })
-                        .doOnError(e -> log.error(
-                            "streamAsk() error during resolved streaming — conversationId={}", srCapturedConvId, e))
-                        .onErrorResume(e -> {
-                            if (!srHasEmitted.get()) {
-                                return Flux.error(e);
-                            }
-                            try {
-                                SseErrorEvent errorEvent = AskStreamErrorMapper.toSseErrorEvent(e);
-                                String json = objectMapper.writeValueAsString(errorEvent);
-                                return Flux.just(json);
-                            } catch (Exception serEx) {
-                                log.error("streamAsk() failed to serialize resolved error event", serEx);
-                                return Flux.empty();
-                            }
-                        }));
-                }
-                // Resolution failed: fall through to normal RAG (cache already cleared)
-                log.info("streamAsk() clarification unresolvable — falling back to normal RAG conversationId={}",
-                    conversationId);
-            }
-        }
-
-        if (conversationId != null && !conversationId.isBlank()) {
-            Optional<ProcedureContextService.ProcedureAmbiguityResult> ambiguity =
-                procedureContextService.detectProcedureAmbiguity(question, cityId);
-            if (ambiguity.isPresent()) {
-                List<ConversationManagementService.ClarificationCandidate> candidates =
-                    ambiguity.get().candidates();
-                conversationService.storePendingClarification(conversationId, candidates);
-                List<String> titles = candidates.stream().map(c -> c.title()).toList();
-                String clarificationMsg = promptBuilder.buildProcedureClarificationMessage(
-                    titles, detectedLanguage, cityId);
-                log.info("streamAsk() procedure ambiguity clarification triggered — conversationId={} candidateCount={}",
-                    conversationId, candidates.size());
-                return buildFallbackStreamResult(clarificationMsg, conversationId, question);
-            }
-        }
-
-        List<Map<String, Object>> messages = new ArrayList<>();
-        messages.add(Map.of("role", "system", "content", promptBuilder.getSystemMessage(cityId, detectedLanguage)));
-
-        RagContexts ragContexts = buildRagContexts(question, cityId, conversationId, "streamAsk()");
-        ProcedureContextResult procCtx = ragContexts.procedureContext();
-        ProcedureContextService.EventContextResult eventCtx = ragContexts.eventContext();
-        NewsContextResult newsCtx = ragContexts.newsContext();
-        CityInfoContextResult cityInfoCtx = ragContexts.cityInfoContext();
-
-        if (ragContexts.newsIntent() && (newsCtx == null || newsCtx.getSources().isEmpty())) {
-            String fallbackMessage = buildNoNewsFoundMessage(detectedLanguage, cityId);
-            log.info("streamAsk() news fallback used - conversationId={} city={}", conversationId, cityId);
-            return buildFallbackStreamResult(fallbackMessage, conversationId, question);
-        }
-
-        if (procCtx != null && procCtx.getContextBlock() != null)
-            messages.add(Map.of("role", "system", "content", procCtx.getContextBlock()));
-        if (eventCtx != null && eventCtx.getContextBlock() != null)
-            messages.add(Map.of("role", "system", "content", eventCtx.getContextBlock()));
-        if (newsCtx != null && newsCtx.getContextBlock() != null)
-            messages.add(Map.of("role", "system", "content", newsCtx.getContextBlock()));
-        if (cityInfoCtx != null && cityInfoCtx.getContextBlock() != null)
-            messages.add(Map.of("role", "system", "content", cityInfoCtx.getContextBlock()));
-
-        var history = conversationService.getConversationHistory(conversationId);
-        conversationService.addToMessages(messages, history);
-        messages.add(Map.of("role", "user", "content", question != null ? question.trim() : ""));
-
-        // 4. Collect sources from RAG context with URL validation
-        List<Source> validatedSources = new ArrayList<>();
-        validateAndAddSources(validatedSources, procCtx != null ? procCtx.getSources() : null, "procedure");
-        validateAndAddSources(validatedSources, eventCtx != null ? eventCtx.getSources() : null, "event");
-        validateAndAddSources(validatedSources, newsCtx != null ? newsCtx.getSources() : null, "news");
-        validateAndAddSources(validatedSources, cityInfoCtx != null ? cityInfoCtx.getSources() : null, "city_info");
-
-        List<SseSources> allSources = validatedSources.stream()
-                .map(s -> new SseSources(s.getTitle(), s.getUrl()))
-                .collect(Collectors.toList());
-
-        // 5. Stream chunks, then emit sources, then done
-        final String capturedQuestion = question;
-        final String capturedCityId = cityId;
-        final String capturedConvId = conversationId;
-        final StringBuilder assembled = new StringBuilder();
-        final java.util.concurrent.atomic.AtomicBoolean hasEmitted = new java.util.concurrent.atomic.AtomicBoolean(
-                false);
-
-        log.debug("streamAsk() preparing to stream — conversationId={} sources={}", capturedConvId, allSources.size());
-
-        // Verify objectMapper is not null before proceeding
-        if (objectMapper == null) {
-            log.error("streamAsk() ERROR: objectMapper is null! This should never happen.");
-            return new AskStreamResult.Error(AskStreamErrorMapper.toResponseDto(
-                    new IllegalStateException("ObjectMapper is not properly injected")));
-        }
-
-        OpenRouterStreamStartResult streamStartResult = httpWrapper.streamFromOpenRouter(messages);
-        if (streamStartResult instanceof OpenRouterStreamStartResult.Error error) {
-            return new AskStreamResult.Error(AskStreamErrorMapper.toResponseDto(error.failure()));
-        }
-
-        Publisher<String> stream = ((OpenRouterStreamStartResult.Success) streamStartResult).stream();
-
-        return new AskStreamResult.Success(Flux.from(stream)
-                .doOnSubscribe(sub -> {
-                }) // Signal subscription start (debugging removed)
-                .handle((raw, sink) -> {
-                    SseChunkParser.ParseResult parsed = SseChunkParser.parseLine(raw);
-                    if (parsed.state() == SseChunkParser.ParseState.MALFORMED) {
-                        sink.error(new IllegalStateException("Malformed upstream stream chunk"));
-                        return;
-                    }
-                    if (parsed.state() == SseChunkParser.ParseState.DELTA && parsed.delta() != null
-                            && !parsed.delta().isEmpty()) {
-                        sink.next(parsed.delta());
-                    }
-                })
-                .cast(String.class)
-                .doOnNext(delta -> {
-                    assembled.append(delta);
-                    hasEmitted.set(true);
-                    log.debug("streamAsk() chunk delta processed");
-                })
-                .map(delta -> {
-                    // Serialize each chunk delta to JSON
-                    try {
-                        SseChunkEvent chunk = new SseChunkEvent(delta);
-                        String serialized = objectMapper.writeValueAsString(chunk);
-                        log.debug("streamAsk() emitting chunk event serialized");
-                        return serialized;
-                    } catch (Exception e) {
-                        log.error("Failed to serialize chunk", e);
-                        throw new RuntimeException("Chunk serialization failed", e);
-                    }
-                })
-                .doOnComplete(() -> {
-                    log.info("First Flux completed");
-                })
-                .concatWith(produceSourcesAndDone(allSources, capturedConvId))
-                .doOnNext(event -> log.debug("streamAsk() emitting event"))
-                .doOnComplete(() -> {
-                    String fullText = assembled.toString();
-                    if (capturedConvId != null && !capturedConvId.isBlank() && !fullText.isBlank()) {
-                        conversationService.updateConversationHistory(capturedConvId, capturedQuestion, fullText);
-                    }
-                    log.info("streamAsk() completed — conversationId={} assembledLength={} city={}",
-                            capturedConvId, fullText.length(), capturedCityId);
-                })
-                .doOnError(e -> {
-                    log.error("streamAsk() error during streaming — conversationId={}", capturedConvId, e);
-                })
-                .onErrorResume(e -> {
-                    // Only convert to SSE error event if data has already been emitted.
-                    // If nothing was emitted yet, let the error propagate so Micronaut
-                    // can return an HTTP 5xx response (no headers committed yet).
-                    if (!hasEmitted.get()) {
-                        return Flux.error(e);
-                    }
-                    try {
-                        SseErrorEvent errorEvent = AskStreamErrorMapper.toSseErrorEvent(e);
-                        String json = objectMapper.writeValueAsString(errorEvent);
-                        return Flux.just(json);
-                    } catch (Exception serEx) {
-                        log.error("streamAsk() failed to serialize streaming error event", serEx);
-                        return Flux.empty();
-                    }
-                }));
+        return streamingOrchestrator.streamAsk(question, conversationId, cityId);
     }
 
-    /**
-     * Produces sources and done events for SSE stream.
-     */
-    private Publisher<String> produceSourcesAndDone(List<SseSources> sources, String conversationId) {
-        try {
-            log.info("produceSourcesAndDone() CALLED — sources={} conversationId={}",
-                    sources != null ? sources.size() : "null", conversationId);
-
-            if (objectMapper == null) {
-                log.error("produceSourcesAndDone() ERROR: objectMapper is null!");
-                return Flux.error(new IllegalStateException("ObjectMapper is null in produceSourcesAndDone"));
-            }
-
-            SseSourcesEvent sourcesEvent = new SseSourcesEvent(sources);
-            final String sourcesJson = objectMapper.writeValueAsString(sourcesEvent);
-
-            SseDoneEvent doneEvent = new SseDoneEvent(conversationId);
-            final String doneJson = objectMapper.writeValueAsString(doneEvent);
-
-            Publisher<String> result = Flux.just(sourcesJson, doneJson);
-            log.info("produceSourcesAndDone() RETURNING Flux with 2 events");
-            return result;
-        } catch (Exception e) {
-            log.error("produceSourcesAndDone() EXCEPTION: Failed to serialize sources or done event", e);
-            return Flux.error(e);
-        }
-    }
+    // -----------------------------------------------------------------------
+    // Private helpers — RAG context building
+    // -----------------------------------------------------------------------
 
     private RagContexts buildRagContexts(String question, String cityId, String conversationId, String operationName) {
         long detectionStart = System.nanoTime();
-        ContextRequirements requirements = procedureContextService.detectContextRequirements(question, cityId);
+        ContextRequirements requirements = intentDetector.detectContextRequirements(question, cityId);
         long detectionDurationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - detectionStart);
 
         log.debug(
@@ -894,13 +363,13 @@ public class OpenRouterServices implements IOpenRouterService {
 
         if (requirements.needsProcedureContext() && requirements.needsEventContext()) {
             long ragStart = System.nanoTime();
-            CompletableFuture<ProcedureContextResult> procedureFuture = procedureContextService
+            CompletableFuture<ProcedureContextResult> procedureFuture = ragContextBuilder
                     .buildProcedureContextResultAsync(question, cityId, ragExecutor)
                     .exceptionally(ex -> {
                         logRagFailure(operationName, conversationId, cityId, "procedure", ex);
                         return null;
                     });
-            CompletableFuture<ProcedureContextService.EventContextResult> eventFuture = procedureContextService
+            CompletableFuture<EventContextResult> eventFuture = ragContextBuilder
                     .buildEventContextResultAsync(question, cityId, ragExecutor)
                     .exceptionally(ex -> {
                         logRagFailure(operationName, conversationId, cityId, "event", ex);
@@ -908,7 +377,7 @@ public class OpenRouterServices implements IOpenRouterService {
                     });
 
             ProcedureContextResult procedureContext = procedureFuture.join();
-            ProcedureContextService.EventContextResult eventContext = eventFuture.join();
+            EventContextResult eventContext = eventFuture.join();
 
             long ragDurationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - ragStart);
             log.debug("{} RAG build completed — conversationId={} mode=bounded-parallel durationMs={}",
@@ -917,31 +386,31 @@ public class OpenRouterServices implements IOpenRouterService {
         }
 
         if (requirements.needsProcedureContext()) {
-            ProcedureContextResult procedureContext = safelyBuildProcedureContext(question, cityId, conversationId,
-                    operationName);
+            ProcedureContextResult procedureContext = safelyBuildProcedureContext(
+                    question, cityId, conversationId, operationName);
             return new RagContexts(procedureContext, null, null, null, false);
         }
 
         if (requirements.needsEventContext()) {
-            ProcedureContextService.EventContextResult eventContext = safelyBuildEventContext(question, cityId,
-                    conversationId, operationName);
+            EventContextResult eventContext = safelyBuildEventContext(
+                    question, cityId, conversationId, operationName);
             return new RagContexts(null, eventContext, null, null, false);
         }
 
         if (requirements.needsCityInfoContext()) {
-            CityInfoContextResult cityInfoContext = safelyBuildCityInfoContext(question, cityId, conversationId,
-                    operationName);
+            CityInfoContextResult cityInfoContext = safelyBuildCityInfoContext(
+                    question, cityId, conversationId, operationName);
             return new RagContexts(null, null, null, cityInfoContext, false);
         }
 
         return new RagContexts(null, null, null, null, false);
     }
 
-    private NewsContextResult safelyBuildNewsContext(String question, String cityId, String conversationId,
-            String operationName) {
+    private NewsContextResult safelyBuildNewsContext(String question, String cityId,
+                                                      String conversationId, String operationName) {
         long start = System.nanoTime();
         try {
-            NewsContextResult result = procedureContextService.buildNewsContextResult(question, cityId);
+            NewsContextResult result = ragContextBuilder.buildNewsContextResult(question, cityId);
             long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - start);
             log.debug("{} RAG build completed — conversationId={} mode=news-only durationMs={}",
                     operationName, conversationId, durationMs);
@@ -952,11 +421,11 @@ public class OpenRouterServices implements IOpenRouterService {
         }
     }
 
-    private ProcedureContextResult safelyBuildProcedureContext(String question, String cityId, String conversationId,
-            String operationName) {
+    private ProcedureContextResult safelyBuildProcedureContext(String question, String cityId,
+                                                                String conversationId, String operationName) {
         long start = System.nanoTime();
         try {
-            ProcedureContextResult result = procedureContextService.buildProcedureContextResult(question, cityId);
+            ProcedureContextResult result = ragContextBuilder.buildProcedureContextResult(question, cityId);
             long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - start);
             log.debug("{} RAG build completed — conversationId={} mode=procedure-only durationMs={}",
                     operationName, conversationId, durationMs);
@@ -967,13 +436,11 @@ public class OpenRouterServices implements IOpenRouterService {
         }
     }
 
-    private ProcedureContextService.EventContextResult safelyBuildEventContext(String question, String cityId,
-            String conversationId, String operationName) {
+    private EventContextResult safelyBuildEventContext(String question, String cityId,
+                                                        String conversationId, String operationName) {
         long start = System.nanoTime();
         try {
-            ProcedureContextService.EventContextResult result = procedureContextService.buildEventContextResult(
-                    question,
-                    cityId);
+            EventContextResult result = ragContextBuilder.buildEventContextResult(question, cityId);
             long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - start);
             log.debug("{} RAG build completed — conversationId={} mode=event-only durationMs={}",
                     operationName, conversationId, durationMs);
@@ -984,11 +451,11 @@ public class OpenRouterServices implements IOpenRouterService {
         }
     }
 
-    private CityInfoContextResult safelyBuildCityInfoContext(String question, String cityId, String conversationId,
-            String operationName) {
+    private CityInfoContextResult safelyBuildCityInfoContext(String question, String cityId,
+                                                              String conversationId, String operationName) {
         long start = System.nanoTime();
         try {
-            CityInfoContextResult result = procedureContextService.buildCityInfoContextResult(question, cityId);
+            CityInfoContextResult result = ragContextBuilder.buildCityInfoContextResult(question, cityId);
             long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - start);
             log.debug("{} RAG build completed — conversationId={} mode=cityinfo-only durationMs={}",
                     operationName, conversationId, durationMs);
@@ -999,14 +466,18 @@ public class OpenRouterServices implements IOpenRouterService {
         }
     }
 
-    private void logRagFailure(String operationName, String conversationId, String cityId, String contextType,
-            Throwable failure) {
+    private void logRagFailure(String operationName, String conversationId, String cityId,
+                                String contextType, Throwable failure) {
         Throwable rootCause = failure instanceof CompletionException && failure.getCause() != null
                 ? failure.getCause()
                 : failure;
         log.warn("{} RAG build failed — conversationId={} city={} contextType={} error={}",
                 operationName, conversationId, cityId, contextType, rootCause.getMessage(), rootCause);
     }
+
+    // -----------------------------------------------------------------------
+    // Private helpers — utilities
+    // -----------------------------------------------------------------------
 
     private ExecutorService createRagExecutor() {
         return Executors.newFixedThreadPool(2, runnable -> {
@@ -1016,25 +487,27 @@ public class OpenRouterServices implements IOpenRouterService {
         });
     }
 
-    private AskStreamResult buildFallbackStreamResult(String fallbackMessage, String conversationId, String question) {
-        try {
-            SseChunkEvent chunk = new SseChunkEvent(fallbackMessage);
-            SseSourcesEvent sourcesEvent = new SseSourcesEvent(List.of());
-            SseDoneEvent doneEvent = new SseDoneEvent(conversationId);
-
-            String chunkJson = objectMapper.writeValueAsString(chunk);
-            String sourcesJson = objectMapper.writeValueAsString(sourcesEvent);
-            String doneJson = objectMapper.writeValueAsString(doneEvent);
-
-            if (conversationId != null && !conversationId.isBlank()) {
-                conversationService.updateConversationHistory(conversationId, question, fallbackMessage);
-            }
-
-            return new AskStreamResult.Success(Flux.just(chunkJson, sourcesJson, doneJson));
-        } catch (Exception e) {
-            log.error("streamAsk() failed to serialize fallback news response", e);
-            return new AskStreamResult.Error(AskStreamErrorMapper.toResponseDto(e));
+    private static int estimateTokenCount(String text) {
+        if (text == null || text.isEmpty()) {
+            return 0;
         }
+        return Math.max(1, text.length() / 4);
+    }
+
+    private static long computeSourcesHash(List<Source> sources) {
+        if (sources == null || sources.isEmpty()) {
+            return 0;
+        }
+        StringBuilder sb = new StringBuilder();
+        for (Source source : sources) {
+            String title = source.getTitle();
+            String url = source.getUrl();
+            if (title != null)
+                sb.append(title).append("|");
+            if (url != null)
+                sb.append(url).append("|");
+        }
+        return sb.toString().hashCode();
     }
 
     private String buildNoNewsFoundMessage(String detectedLanguage, String cityId) {
@@ -1059,10 +532,9 @@ public class OpenRouterServices implements IOpenRouterService {
     }
 
     private record RagContexts(ProcedureContextResult procedureContext,
-            ProcedureContextService.EventContextResult eventContext,
-            NewsContextResult newsContext,
-            CityInfoContextResult cityInfoContext,
-            boolean newsIntent) {
+                               EventContextResult eventContext,
+                               NewsContextResult newsContext,
+                               CityInfoContextResult cityInfoContext,
+                               boolean newsIntent) {
     }
-
 }
