--- src/main/java/cat/complai/services/openrouter/procedure/ProcedureContextService.java
@@ -1,1109 +1,0 @@
-package cat.complai.services.openrouter.procedure;
-
-import cat.complai.dto.openrouter.Source;
-import cat.complai.helpers.openrouter.CityInfoRagHelperRegistry;
-import cat.complai.helpers.openrouter.EventRagHelperRegistry;
-import cat.complai.helpers.openrouter.GencatProcedureRagHelperRegistry;
-import cat.complai.helpers.openrouter.NewsRagHelperRegistry;
-import cat.complai.helpers.openrouter.ProcedureRagHelperRegistry;
-import cat.complai.helpers.openrouter.RagHelper;
-import cat.complai.helpers.openrouter.RedactPromptBuilder;
-import cat.complai.helpers.openrouter.rag.AmbiguityDetector;
-import cat.complai.helpers.openrouter.rag.InMemoryLexicalIndex;
-import cat.complai.helpers.openrouter.rag.RagJavaCalibration;
-import cat.complai.helpers.openrouter.rag.TokenNormalizer;
-import cat.complai.services.openrouter.conversation.ConversationManagementService;
-import jakarta.inject.Inject;
-import jakarta.inject.Singleton;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.Executor;
-import java.util.concurrent.TimeUnit;
-import java.util.regex.Pattern;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
-/**
- * Retrieves RAG context blocks for procedures, events, news, and city
- * information
- * to inject into AI prompts.
- *
- * <p>
- * Each domain (procedures, events, news, city-info) is served by its own helper
- * registry
- * that lazily builds an in-memory lexical index per city on first access. This
- * service
- * orchestrates keyword-based intent detection and context extraction, then
- * returns typed
- * result objects that carry both the context text and the source URLs for
- * attribution.
- */
-@Singleton
-public class ProcedureContextService {
-
-    private static final Set<String> TITLE_STOP_WORDS = Set.of(
-            "a", "an", "and", "d", "de", "del", "el", "els", "en", "for", "i", "la", "las",
-            "les", "lo", "los", "of", "the", "y");
-
-    private static final List<String> PROCEDURE_KEYWORDS = List.of(
-            "how to", "how do i", "what is the process", "procedure", "tramit", "tràmit",
-            "requirement", "requirements", "document", "documents", "apply", "application",
-            "form", "forms", "permit", "license", "request", "complaint", "claim",
-            "where can i", "how can i", "steps", "step by step", "process",
-            "recycling", "waste", "garbage", "trash", "center", "collection");
-
-    private static final List<String> GENERALITAT_KEYWORDS = List.of(
-            "generalitat", "generalitat de catalunya", "catalan government",
-            "gencat", "gencat procediment", "tramit gencat",
-            "govern catala", "catalunya", "catalan");
-
-    private static final List<String> MUNICIPAL_SERVICE_KEYWORDS = List.of(
-            "ajuntament", "city hall", "municipal", "council");
-
-    private static final List<String> EVENT_KEYWORDS = List.of(
-            "event", "events", "evento", "eventos", "esdeveniment", "esdeveniments",
-            "agenda", "activity", "activities", "activitat", "activitats",
-            "festival", "concert", "exhibition", "exposició", "theater", "teatre",
-            "cinema", "sports", "esports", "culture", "cultura", "celebration", "celebració",
-            "what's on", "what's happening", "què passa", "agenda cultural", "program",
-            "this weekend", "upcoming",
-            "capsa", "artesa", "l'artesa", "la capsa", "teatre artesa");
-
-    private static final List<String> EVENT_INTENT_GUARD_KEYWORDS = List.of(
-            "event", "events", "evento", "eventos", "esdeveniment", "esdeveniments",
-            "agenda", "activity", "activities", "activitat", "activitats",
-            "what's on", "what is on", "what's happening", "que passa", "què passa");
-
-    private static final List<String> DATE_WINDOW_KEYWORDS = List.of(
-            "today", "tomorrow", "tonight", "this weekend", "next weekend", "weekend",
-            "this week", "next week", "this month", "next month",
-            "avui", "dema", "demà", "aquesta setmana", "setmana vinent", "cap de setmana", "aquest mes",
-            "mes vinent",
-            "hoy", "manana", "mañana", "esta semana", "proxima semana", "próxima semana", "fin de semana",
-            "este mes", "proximo mes", "próximo mes");
-
-    private static final List<String> MONTH_KEYWORDS = List.of(
-            "january", "february", "march", "april", "may", "june", "july", "august",
-            "september", "october", "november", "december",
-            "gener", "febrer", "marc", "març", "abril", "maig", "juny", "juliol", "agost",
-            "setembre", "octubre", "novembre", "desembre",
-            "enero", "febrero", "marzo", "abril", "mayo", "junio", "julio", "agosto",
-            "septiembre", "octubre", "noviembre", "diciembre");
-
-    private static final List<String> RANGE_CONNECTOR_KEYWORDS = List.of(
-            "from", "to", "between", "until", "through", "del", "al", "desde", "hasta", "entre");
-
-    private static final Pattern NUMERIC_DATE_PATTERN = Pattern.compile("\\b\\d{1,2}[/-]\\d{1,2}(?:[/-]\\d{2,4})?\\b");
-    private static final Pattern ISO_DATE_PATTERN = Pattern.compile("\\b\\d{4}-\\d{1,2}-\\d{1,2}\\b");
-
-    private static final List<String> NEWS_KEYWORDS = List.of(
-            "news", "latest news", "recent news", "recent happenings", "current affairs",
-            "actuality", "actualitat", "actualidad", "noticies", "notícies", "noticias",
-            "ultimes", "últimes", "ultimas", "últimas", "latest", "recent");
-
-    private static final List<String> CITY_INFO_KEYWORDS = List.of(
-            "city", "municipal", "ajuntament", "city hall", "serveis", "serveis municipals",
-            "services", "public service", "public services", "equipaments", "equipamientos",
-            "tourism", "turisme", "turismo", "visit", "visitar", "visita", "informacio",
-            "informació", "informacion", "información");
-
-    private static final List<String> CONVERSATIONAL_SHORT_CIRCUITS = List.of(
-            "hello", "hi", "hey", "good morning", "good afternoon", "good evening",
-            "hola", "bon dia", "bona tarda", "how are you", "what is the weather", "weather",
-            "tell me a joke", "who are you", "what's your name", "thank you", "thanks");
-
-    private final ProcedureRagHelperRegistry ragRegistry;
-    private final GencatProcedureRagHelperRegistry gencatRagRegistry;
-    private final EventRagHelperRegistry eventRagRegistry;
-    private final NewsRagHelperRegistry newsRagRegistry;
-    private final CityInfoRagHelperRegistry cityInfoRagRegistry;
-    @SuppressWarnings("unused")
-    private final RedactPromptBuilder promptBuilder;
-    private final Logger logger = Logger.getLogger(ProcedureContextService.class.getName());
-    private final ConcurrentHashMap<String, CityDetectionIndex> detectionIndexByCity = new ConcurrentHashMap<>();
-
-    /**
-     * Constructs the service with all RAG helper registries and the prompt builder.
-     *
-     * @param ragRegistry         registry for procedure RAG helpers
-     * @param gencatRagRegistry registry for gencat procedure RAG helpers (fallback)
-     * @param eventRagRegistry    registry for event RAG helpers
-     * @param newsRagRegistry     registry for news RAG helpers
-     * @param cityInfoRagRegistry registry for city-info RAG helpers
-     * @param promptBuilder       builder used to resolve procedure keywords per
-     *                            city
-     */
-    @Inject
-    public ProcedureContextService(ProcedureRagHelperRegistry ragRegistry,
-            GencatProcedureRagHelperRegistry gencatRagRegistry,
-            EventRagHelperRegistry eventRagRegistry,
-            NewsRagHelperRegistry newsRagRegistry,
-            CityInfoRagHelperRegistry cityInfoRagRegistry,
-            RedactPromptBuilder promptBuilder) {
-        this.ragRegistry = ragRegistry;
-        this.gencatRagRegistry = gencatRagRegistry;
-        this.eventRagRegistry = eventRagRegistry;
-        this.newsRagRegistry = newsRagRegistry;
-        this.cityInfoRagRegistry = cityInfoRagRegistry;
-        this.promptBuilder = promptBuilder;
-    }
-
-    /**
-     * Convenience constructor with all registries (used by tests).
-     */
-    public ProcedureContextService(ProcedureRagHelperRegistry ragRegistry,
-            EventRagHelperRegistry eventRagRegistry,
-            NewsRagHelperRegistry newsRagRegistry,
-            CityInfoRagHelperRegistry cityInfoRagRegistry,
-            RedactPromptBuilder promptBuilder) {
-        this(ragRegistry, new GencatProcedureRagHelperRegistry(), eventRagRegistry, newsRagRegistry,
-                cityInfoRagRegistry, promptBuilder);
-    }
-
-    /**
-     * Convenience constructor (used by tests).
-     */
-    public ProcedureContextService(ProcedureRagHelperRegistry ragRegistry, EventRagHelperRegistry eventRagRegistry,
-            RedactPromptBuilder promptBuilder) {
-        this(ragRegistry, new GencatProcedureRagHelperRegistry(), eventRagRegistry, new NewsRagHelperRegistry(),
-                new CityInfoRagHelperRegistry(), promptBuilder);
-    }
-
-    /**
-     * Result type for procedure context extraction: context block string and the
-     * list of source URLs.
-     */
-    public static class ProcedureContextResult {
-        private final String contextBlock;
-        private final List<Source> sources;
-
-        public ProcedureContextResult(String contextBlock, List<Source> sources) {
-            this.contextBlock = contextBlock;
-            this.sources = sources == null ? List.of() : Collections.unmodifiableList(new ArrayList<>(sources));
-        }
-
-        public String getContextBlock() {
-            return contextBlock;
-        }
-
-        public List<Source> getSources() {
-            return sources;
-        }
-    }
-
-    /**
-     * Result type for event context extraction: context block string and the list
-     * of source URLs.
-     */
-    public static class EventContextResult {
-        private final String contextBlock;
-        private final List<Source> sources;
-
-        public EventContextResult(String contextBlock, List<Source> sources) {
-            this.contextBlock = contextBlock;
-            this.sources = sources == null ? List.of() : Collections.unmodifiableList(new ArrayList<>(sources));
-        }
-
-        public String getContextBlock() {
-            return contextBlock;
-        }
-
-        public List<Source> getSources() {
-            return sources;
-        }
-    }
-
-    /**
-     * Result type for news context extraction.
-     */
-    public static class NewsContextResult {
-        private final String contextBlock;
-        private final List<Source> sources;
-
-        /**
-         * Constructs a news context result.
-         *
-         * @param contextBlock the RAG context text to inject into the AI prompt
-         * @param sources      the source documents used to build the context
-         */
-        public NewsContextResult(String contextBlock, List<Source> sources) {
-            this.contextBlock = contextBlock;
-            this.sources = sources == null ? List.of() : Collections.unmodifiableList(new ArrayList<>(sources));
-        }
-
-        /**
-         * Returns the RAG context text snippet for the AI prompt.
-         *
-         * @return context block text
-         */
-        public String getContextBlock() {
-            return contextBlock;
-        }
-
-        /**
-         * Returns the unmodifiable list of source documents.
-         *
-         * @return list of sources
-         */
-        public List<Source> getSources() {
-            return sources;
-        }
-    }
-
-    /**
-     * Result type for city-information context extraction.
-     */
-    public static class CityInfoContextResult {
-        private final String contextBlock;
-        private final List<Source> sources;
-
-        /**
-         * Constructs a city-info context result.
-         *
-         * @param contextBlock the RAG context text to inject into the AI prompt
-         * @param sources      the source documents used to build the context
-         */
-        public CityInfoContextResult(String contextBlock, List<Source> sources) {
-            this.contextBlock = contextBlock;
-            this.sources = sources == null ? List.of() : Collections.unmodifiableList(new ArrayList<>(sources));
-        }
-
-        /**
-         * Returns the RAG context text snippet for the AI prompt.
-         *
-         * @return context block text
-         */
-        public String getContextBlock() {
-            return contextBlock;
-        }
-
-        /**
-         * Returns the unmodifiable list of source documents.
-         *
-         * @return list of sources
-         */
-        public List<Source> getSources() {
-            return sources;
-        }
-    }
-
-    public record ProcedureAmbiguityResult(
-            List<ConversationManagementService.ClarificationCandidate> candidates) {}
-
-    /**
-     * Captures which RAG context domains are required for a given query.
-     */
-    public static final class ContextRequirements {
-        private static final ContextRequirements NONE = new ContextRequirements(false, false, false, false);
-
-        private final boolean needsProcedureContext;
-        private final boolean needsEventContext;
-        private final boolean needsNewsContext;
-        private final boolean needsCityInfoContext;
-
-        /**
-         * Constructs a requirements descriptor.
-         *
-         * @param needsProcedureContext    {@code true} if procedure context should be
-         *                                 fetched
-         * @param needsEventContext        {@code true} if event context should be
-         *                                 fetched
-         * @param needsNewsContext         {@code true} if news context should be
-         *                                 fetched
-         * @param needsCityInfoContext     {@code true} if city-info context should be
-         *                                 fetched
-         */
-        public ContextRequirements(boolean needsProcedureContext, boolean needsEventContext, boolean needsNewsContext,
-                boolean needsCityInfoContext) {
-            this.needsProcedureContext = needsProcedureContext;
-            this.needsEventContext = needsEventContext;
-            this.needsNewsContext = needsNewsContext;
-            this.needsCityInfoContext = needsCityInfoContext;
-        }
-
-        /**
-         * Returns an instance that requires no context at all.
-         *
-         * @return shared NONE instance
-         */
-        public static ContextRequirements none() {
-            return NONE;
-        }
-
-        /**
-         * Returns {@code true} if procedure context should be fetched.
-         *
-         * @return needs-procedure flag
-         */
-        public boolean needsProcedureContext() {
-            return needsProcedureContext;
-        }
-
-        /**
-         * Returns {@code true} if event context should be fetched.
-         *
-         * @return needs-event flag
-         */
-        public boolean needsEventContext() {
-            return needsEventContext;
-        }
-
-        /**
-         * Returns {@code true} if news context should be fetched.
-         *
-         * @return needs-news flag
-         */
-        public boolean needsNewsContext() {
-            return needsNewsContext;
-        }
-
-        /**
-         * Returns {@code true} if city-info context should be fetched.
-         *
-         * @return needs-city-info flag
-         */
-        public boolean needsCityInfoContext() {
-            return needsCityInfoContext;
-        }
-    }
-
-    private record CityDetectionIndex(TitleMatcher procedureMatcher, TitleMatcher eventMatcher,
-            TitleMatcher newsMatcher, TitleMatcher cityInfoMatcher,
-            int procedureTitleCount, int eventTitleCount, int newsTitleCount, int cityInfoTitleCount) {
-    }
-
-    private static final class TitleMatcher {
-        private final Map<String, List<String>> titlesByToken;
-        private final List<String> fallbackTitles;
-
-        private TitleMatcher(Map<String, List<String>> titlesByToken, List<String> fallbackTitles) {
-            this.titlesByToken = titlesByToken;
-            this.fallbackTitles = fallbackTitles;
-        }
-
-        private static TitleMatcher fromTitles(List<String> normalizedTitles) {
-            Map<String, LinkedHashSet<String>> byToken = new LinkedHashMap<>();
-            List<String> fallbackTitles = new ArrayList<>();
-
-            for (String normalizedTitle : normalizedTitles) {
-                Set<String> tokens = tokenize(normalizedTitle);
-                if (tokens.isEmpty()) {
-                    fallbackTitles.add(normalizedTitle);
-                    continue;
-                }
-                for (String token : tokens) {
-                    byToken.computeIfAbsent(token, ignored -> new LinkedHashSet<>()).add(normalizedTitle);
-                }
-            }
-
-            Map<String, List<String>> immutableIndex = new LinkedHashMap<>();
-            byToken.forEach((token, titles) -> immutableIndex.put(token, List.copyOf(titles)));
-            return new TitleMatcher(Map.copyOf(immutableIndex), List.copyOf(fallbackTitles));
-        }
-
-        private boolean matches(String normalizedQuestion) {
-            if (normalizedQuestion == null || normalizedQuestion.isBlank()) {
-                return false;
-            }
-
-            LinkedHashSet<String> candidates = new LinkedHashSet<>();
-            for (String token : tokenize(normalizedQuestion)) {
-                List<String> titles = titlesByToken.get(token);
-                if (titles != null) {
-                    candidates.addAll(titles);
-                }
-            }
-
-            if (candidates.isEmpty()) {
-                candidates.addAll(fallbackTitles);
-            }
-
-            for (String candidate : candidates) {
-                if (normalizedQuestion.contains(candidate)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    }
-
-    /**
-     * Determines if a question likely needs procedure/municipal information.
-     * This avoids expensive RAG searches for conversational queries.
-     * Checks against actual procedure titles for more accurate detection.
-     */
-    public boolean questionNeedsProcedureContext(String question, String cityId) {
-        return detectContextRequirements(question, cityId).needsProcedureContext();
-    }
-
-    public ContextRequirements detectContextRequirements(String question, String cityId) {
-        if (question == null || question.isBlank()) {
-            return ContextRequirements.none();
-        }
-
-        String normalizedQuestion = normalize(question);
-        long startNanos = System.nanoTime();
-
-        boolean needsProcedure = containsAny(normalizedQuestion, PROCEDURE_KEYWORDS)
-                || containsAny(normalizedQuestion, MUNICIPAL_SERVICE_KEYWORDS);
-        boolean needsEvent = containsAny(normalizedQuestion, EVENT_KEYWORDS);
-        boolean needsNews = containsAny(normalizedQuestion, NEWS_KEYWORDS);
-
-        if (needsNews) {
-            long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos);
-            logger.fine(() -> "Context detection completed from news keyword path - city=" + cityId
-                    + " procedure=false event=false news=true durationMs=" + durationMs);
-            return new ContextRequirements(false, false, true, false);
-        }
-
-        if (!needsProcedure && !needsEvent && isClearlyConversational(normalizedQuestion)) {
-            logger.fine(() -> "Context detection short-circuited conversational query — city=" + cityId);
-            return ContextRequirements.none();
-        }
-
-        if (needsProcedure ^ needsEvent) {
-            boolean finalNeedsProcedure = needsProcedure;
-            boolean finalNeedsEvent = needsEvent;
-            long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos);
-            logger.fine(() -> "Context detection short-circuited keyword query — city=" + cityId
-                    + " procedure=" + finalNeedsProcedure
-                    + " event=" + finalNeedsEvent
-                    + " news=false"
-                    + " durationMs=" + durationMs);
-            return new ContextRequirements(finalNeedsProcedure, finalNeedsEvent, false, false);
-        }
-
-        if (needsProcedure && needsEvent) {
-            long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos);
-            logger.fine(() -> "Context detection completed from keyword path — city=" + cityId
-                    + " procedure=true event=true news=false durationMs=" + durationMs);
-            return new ContextRequirements(true, true, false, false);
-        }
-
-        CityDetectionIndex detectionIndex = getOrCreateDetectionIndex(cityId);
-        needsProcedure = detectionIndex.procedureMatcher().matches(normalizedQuestion);
-        needsEvent = detectionIndex.eventMatcher().matches(normalizedQuestion);
-        needsNews = detectionIndex.newsMatcher().matches(normalizedQuestion);
-
-        if (needsNews) {
-            long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos);
-            logger.fine(() -> "Context detection completed from title path - city=" + cityId
-                    + " procedure=false event=false news=true durationMs=" + durationMs
-                    + " newsTitles=" + detectionIndex.newsTitleCount());
-            return new ContextRequirements(false, false, true, false);
-        }
-
-        boolean needsCityInfo = false;
-        if (!needsProcedure && !needsEvent && !needsNews) {
-            needsCityInfo = containsAny(normalizedQuestion, CITY_INFO_KEYWORDS)
-                    || detectionIndex.cityInfoMatcher().matches(normalizedQuestion);
-        }
-
-        long durationMs = TimeUnit.NANOSECONDS.toMillis(System.nanoTime() - startNanos);
-        boolean finalNeedsProcedure = needsProcedure;
-        boolean finalNeedsEvent = needsEvent;
-        boolean finalNeedsCityInfo = needsCityInfo;
-        logger.fine(() -> "Context detection completed — city=" + cityId
-                + " procedure=" + finalNeedsProcedure
-                + " event=" + finalNeedsEvent
-                + " news=false"
-                + " cityInfo=" + finalNeedsCityInfo
-                + " durationMs=" + durationMs
-                + " procedureTitles=" + detectionIndex.procedureTitleCount()
-                + " eventTitles=" + detectionIndex.eventTitleCount()
-                + " cityInfoTitles=" + detectionIndex.cityInfoTitleCount());
-        return new ContextRequirements(finalNeedsProcedure, finalNeedsEvent, false, finalNeedsCityInfo);
-    }
-
-    /**
-     * Checks if the query is asking about Generalitat procedures.
-     *
-     * @param question the user's question
-     * @return true if the question mentions generalitat
-     */
-    public boolean isAskingGencatProcedures(String question) {
-        if (question == null || question.isBlank()) {
-            return false;
-        }
-        String normalized = question.toLowerCase(Locale.ROOT);
-        return containsAny(normalized, GENERALITAT_KEYWORDS)
-                || normalized.contains("generalitat")
-                || normalized.contains("gencat")
-                || normalized.contains("catalan government");
-    }
-
-    public ProcedureContextResult buildProcedureContextResult(String query, String cityId) {
-        return buildProcedureContextResult(query, cityId, false);
-    }
-
-    /**
-     * Builds procedure context result with optional fallback to gencat procedures.
-     *
-     * @param query      the search query
-     * @param cityId    the city identifier
-     * @param forceGencat whether to force using gencat procedures instead of city-specific
-     * @return procedure context result
-     */
-    public ProcedureContextResult buildProcedureContextResult(String query, String cityId, boolean forceGencat) {
-        try {
-            // Determine which registry to use
-            RagHelper<RagHelper.Procedure> helper;
-            String effectiveCityId;
-            boolean useGencat = forceGencat || isAskingGencatProcedures(query);
-
-            if (useGencat) {
-                helper = gencatRagRegistry.getForCity("gencat");
-                effectiveCityId = "gencat";
-                logger.info("Using gencat fallback for procedure query: " + query);
-            } else {
-                helper = ragRegistry.getForCity(cityId);
-                effectiveCityId = cityId;
-            }
-
-            List<RagHelper.Procedure> matches = helper.search(query);
-
-            // If no matches from city procedures and not already using gencat, try fallback
-            if (matches.isEmpty() && !useGencat) {
-                logger.info("No city procedure matches for query: " + query + ", trying gencat fallback");
-                return buildProcedureContextResult(query, cityId, true);
-            }
-
-            if (matches.isEmpty()) {
-                return new ProcedureContextResult(null, List.of());
-            }
-
-            List<Source> sources = matches.stream()
-                    .map(p -> new Source(p.url, p.title))
-                    .filter(source -> source.getUrl() != null && !source.getUrl().isBlank())
-                    .toList();
-            String contextBlock = promptBuilder.buildProcedureContextBlockFromMatches(matches, effectiveCityId);
-            return new ProcedureContextResult(contextBlock, sources);
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "Failed to build procedure context result for city=" + cityId
-                    + "; returning empty context: " + e.getMessage(), e);
-            return new ProcedureContextResult(null, List.of());
-        }
-    }
-
-    public Optional<ProcedureAmbiguityResult> detectProcedureAmbiguity(String question, String cityId) {
-        try {
-            if (question == null || question.isBlank()) {
-                return Optional.empty();
-            }
-            if (!detectContextRequirements(question, cityId).needsProcedureContext()) {
-                return Optional.empty();
-            }
-            InMemoryLexicalIndex.SearchResponse<RagHelper.Procedure> response =
-                    ragRegistry.getForCity(cityId).searchWithScores(question);
-            if (!AmbiguityDetector.isAmbiguous(response, RagJavaCalibration.procedure().absoluteFloor())) {
-                return Optional.empty();
-            }
-            List<ConversationManagementService.ClarificationCandidate> candidates =
-                    AmbiguityDetector.getTopCandidates(response, 3).stream()
-                            .filter(sr -> isCandidateRelevantToQuery(sr.source(), question))
-                            .map(sr -> new ConversationManagementService.ClarificationCandidate(
-                                    sr.source().procedureId, sr.source().title))
-                            .toList();
-            if (candidates.size() < 2) {
-                return Optional.empty();
-            }
-            return Optional.of(new ProcedureAmbiguityResult(candidates));
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "detectProcedureAmbiguity failed for city=" + cityId
-                    + ": " + e.getMessage(), e);
-            return Optional.empty();
-        }
-    }
-
-    public ProcedureContextResult buildProcedureContextResultForId(String procedureId, String cityId) {
-        try {
-            RagHelper.Procedure procedure = ragRegistry.getForCity(cityId).getAll()
-                    .stream()
-                    .filter(p -> procedureId.equals(p.procedureId))
-                    .findFirst()
-                    .orElse(null);
-            if (procedure == null) {
-                return new ProcedureContextResult(null, List.of());
-            }
-            String contextBlock = promptBuilder.buildProcedureContextBlockFromMatches(List.of(procedure), cityId);
-            List<Source> sources = List.of(new Source(procedure.url, procedure.title));
-            return new ProcedureContextResult(contextBlock, sources);
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "buildProcedureContextResultForId failed for procedureId=" + procedureId
-                    + " city=" + cityId + ": " + e.getMessage(), e);
-            return new ProcedureContextResult(null, List.of());
-        }
-    }
-
-    /**
-     * Determines if a question likely needs event information.
-     * This avoids expensive RAG searches for conversational queries.
-     * Checks against actual event titles for more accurate detection.
-     */
-    public boolean questionNeedsEventContext(String question, String cityId) {
-        return detectContextRequirements(question, cityId).needsEventContext();
-    }
-
-    public EventContextResult buildEventContextResult(String query, String cityId) {
-        try {
-            RagHelper<RagHelper.Event> helper = eventRagRegistry.getForCity(cityId);
-            List<RagHelper.Event> matches = helper.search(query);
-            if (matches.isEmpty()) {
-                return new EventContextResult(null, List.of());
-            }
-            List<Source> sources = matches.stream()
-                    .map(e -> new Source(e.url, e.title))
-                    .filter(source -> source.getUrl() != null && !source.getUrl().isBlank())
-                    .toList();
-            String contextBlock = buildEventContextBlockFromMatches(matches, cityId);
-            return new EventContextResult(contextBlock, sources);
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "Failed to build event context result for city=" + cityId
-                    + "; returning empty context: " + e.getMessage(), e);
-            return new EventContextResult(null, List.of());
-        }
-    }
-
-    public boolean questionNeedsNewsContext(String question, String cityId) {
-        return detectContextRequirements(question, cityId).needsNewsContext();
-    }
-
-    public boolean questionNeedsCityInfoContext(String question, String cityId) {
-        return detectContextRequirements(question, cityId).needsCityInfoContext();
-    }
-
-    public boolean requiresEventDateWindowClarification(String question, String cityId) {
-        if (question == null || question.isBlank()) {
-            return false;
-        }
-
-        ContextRequirements requirements = detectContextRequirements(question, cityId);
-        if (!requirements.needsEventContext() || requirements.needsNewsContext()) {
-            return false;
-        }
-
-        String normalizedQuestion = normalize(question);
-        if (!containsAny(normalizedQuestion, EVENT_INTENT_GUARD_KEYWORDS)) {
-            return false;
-        }
-
-        return !hasDateWindow(normalizedQuestion, question);
-    }
-
-    public NewsContextResult buildNewsContextResult(String query, String cityId) {
-        try {
-            RagHelper<RagHelper.News> helper = newsRagRegistry.getForCity(cityId);
-            List<RagHelper.News> matches = helper.search(query);
-            if (matches.isEmpty()) {
-                return new NewsContextResult(null, List.of());
-            }
-
-            List<Source> sources = matches.stream()
-                    .map(item -> new Source(item.url, item.title))
-                    .filter(source -> source.getUrl() != null && !source.getUrl().isBlank())
-                    .toList();
-
-            String contextBlock = buildNewsContextBlockFromMatches(matches, cityId);
-            return new NewsContextResult(contextBlock, sources);
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "Failed to build news context result for city=" + cityId
-                    + "; returning empty context: " + e.getMessage(), e);
-            return new NewsContextResult(null, List.of());
-        }
-    }
-
-    public CityInfoContextResult buildCityInfoContextResult(String query, String cityId) {
-        try {
-            RagHelper<RagHelper.CityInfo> helper = cityInfoRagRegistry.getForCity(cityId);
-            List<RagHelper.CityInfo> matches = helper.search(query);
-            if (matches.isEmpty()) {
-                return new CityInfoContextResult(null, List.of());
-            }
-
-            List<Source> sources = matches.stream()
-                    .map(item -> new Source(item.url, item.title))
-                    .filter(source -> source.getUrl() != null && !source.getUrl().isBlank())
-                    .toList();
-
-            String contextBlock = buildCityInfoContextBlockFromMatches(matches, cityId);
-            return new CityInfoContextResult(contextBlock, sources);
-        } catch (Exception e) {
-            logger.log(Level.WARNING, "Failed to build city-info context result for city=" + cityId
-                    + "; returning empty context: " + e.getMessage(), e);
-            return new CityInfoContextResult(null, List.of());
-        }
-    }
-
-    private String buildNewsContextBlockFromMatches(List<RagHelper.News> matches, String cityId) {
-        if (matches.isEmpty()) {
-            return "";
-        }
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("CONTEXT FROM CITY NEWS IN ").append(cityId).append(":\n\n");
-
-        for (int i = 0; i < matches.size(); i++) {
-            RagHelper.News item = matches.get(i);
-            sb.append(i + 1).append(". ").append(item.title).append("\n");
-            if (item.publishedAt != null && !item.publishedAt.isBlank()) {
-                sb.append("   Published: ").append(item.publishedAt).append("\n");
-            }
-            if (item.categories != null && !item.categories.isBlank()) {
-                sb.append("   Categories: ").append(item.categories).append("\n");
-            }
-            if (item.summary != null && !item.summary.isBlank()) {
-                sb.append("   Summary: ").append(item.summary).append("\n");
-            }
-            if (item.body != null && !item.body.isBlank()) {
-                sb.append("   Details: ").append(item.body).append("\n");
-            }
-            if (item.url != null && !item.url.isBlank()) {
-                sb.append("   Source URL: ").append(item.url).append("\n");
-            }
-            sb.append("\n");
-        }
-
-        sb.append("INSTRUCTIONS:\n");
-        sb.append("- Base your answer on the news context above.\n");
-        sb.append("- Do not invent news items or URLs.\n");
-        sb.append("- MANDATORY: If you cite a news item and it has a Source URL, include that URL in your answer.\n");
-        return sb.toString();
-    }
-
-    private String buildEventContextBlockFromMatches(List<RagHelper.Event> matches, String cityId) {
-        if (matches.isEmpty())
-            return "";
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("Events in ").append(cityId).append(":\n\n");
-
-        for (int i = 0; i < matches.size(); i++) {
-            RagHelper.Event event = matches.get(i);
-            sb.append(i + 1).append(". ").append(event.title).append("\n");
-
-            if (event.date != null && !event.date.isBlank()) {
-                sb.append("   Date: ").append(event.date).append("\n");
-            }
-            if (event.time != null && !event.time.isBlank()) {
-                sb.append("   Time: ").append(event.time).append("\n");
-            }
-            if (event.location != null && !event.location.isBlank()) {
-                sb.append("   Location: ").append(event.location).append("\n");
-            }
-            if (event.eventType != null && !event.eventType.isBlank()) {
-                sb.append("   Type: ").append(event.eventType).append("\n");
-            }
-            if (event.targetAudience != null && !event.targetAudience.isBlank()) {
-                sb.append("   Audience: ").append(event.targetAudience).append("\n");
-            }
-            if (event.description != null && !event.description.isBlank()) {
-                sb.append("   Description: ").append(event.description).append("\n");
-            }
-            if (event.theme != null && !event.theme.isBlank()) {
-                sb.append("   Theme: ").append(event.theme).append("\n");
-            }
-            if (event.url != null && !event.url.isBlank()) {
-                sb.append("   Source URL: ").append(event.url).append("\n");
-            }
-            sb.append("\n");
-        }
-
-        sb.append("INSTRUCTIONS:\n");
-        sb.append("- Base your answer on the events above.\n");
-        sb.append(
-                "- MANDATORY: If you mention an event listed above and it has a Source URL, include that URL in your answer.\n");
-        sb.append("- NEVER invent event URLs.\n");
-
-        return sb.toString();
-    }
-
-    private String buildCityInfoContextBlockFromMatches(List<RagHelper.CityInfo> matches, String cityId) {
-        if (matches.isEmpty()) {
-            return "";
-        }
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("CITY INFORMATION IN ").append(cityId).append(":\n\n");
-
-        for (int i = 0; i < matches.size(); i++) {
-            RagHelper.CityInfo item = matches.get(i);
-            sb.append(i + 1).append(". ").append(item.title).append("\n");
-            if (item.theme != null && !item.theme.isBlank()) {
-                sb.append("   Theme: ").append(item.theme).append("\n");
-            }
-            if (item.breadcrumbs != null && !item.breadcrumbs.isBlank()) {
-                sb.append("   Breadcrumbs: ").append(item.breadcrumbs).append("\n");
-            }
-            if (item.summary != null && !item.summary.isBlank()) {
-                sb.append("   Summary: ").append(item.summary).append("\n");
-            }
-            if (item.body != null && !item.body.isBlank()) {
-                sb.append("   Details: ").append(item.body).append("\n");
-            }
-            if (item.url != null && !item.url.isBlank()) {
-                sb.append("   Source URL: ").append(item.url).append("\n");
-            }
-            sb.append("\n");
-        }
-
-        sb.append("INSTRUCTIONS:\n");
-        sb.append("- Base your answer on the city-info context above.\n");
-        sb.append("- Do not invent city-info items or URLs.\n");
-        sb.append("- MANDATORY: If you cite an item and it has a Source URL, include that URL in your answer.\n");
-        return sb.toString();
-    }
-
-    /**
-     * Async variant of buildProcedureContextResult.
-     * Uses the supplied executor when bounded parallelism is required.
-     */
-    public CompletableFuture<ProcedureContextResult> buildProcedureContextResultAsync(String query, String cityId,
-            Executor executor) {
-        return CompletableFuture.supplyAsync(() -> buildProcedureContextResult(query, cityId), executor);
-    }
-
-    public CompletableFuture<ProcedureContextResult> buildProcedureContextResultAsync(String query, String cityId) {
-        return CompletableFuture.completedFuture(buildProcedureContextResult(query, cityId));
-    }
-
-    /**
-     * Async variant of buildEventContextResult.
-     * Uses the supplied executor when bounded parallelism is required.
-     */
-    public CompletableFuture<EventContextResult> buildEventContextResultAsync(String query, String cityId,
-            Executor executor) {
-        return CompletableFuture.supplyAsync(() -> buildEventContextResult(query, cityId), executor);
-    }
-
-    public CompletableFuture<EventContextResult> buildEventContextResultAsync(String query, String cityId) {
-        return CompletableFuture.completedFuture(buildEventContextResult(query, cityId));
-    }
-
-    public CompletableFuture<NewsContextResult> buildNewsContextResultAsync(String query, String cityId,
-            Executor executor) {
-        return CompletableFuture.supplyAsync(() -> buildNewsContextResult(query, cityId), executor);
-    }
-
-    public CompletableFuture<NewsContextResult> buildNewsContextResultAsync(String query, String cityId) {
-        return CompletableFuture.completedFuture(buildNewsContextResult(query, cityId));
-    }
-
-    public CompletableFuture<CityInfoContextResult> buildCityInfoContextResultAsync(String query, String cityId,
-            Executor executor) {
-        return CompletableFuture.supplyAsync(() -> buildCityInfoContextResult(query, cityId), executor);
-    }
-
-    public CompletableFuture<CityInfoContextResult> buildCityInfoContextResultAsync(String query, String cityId) {
-        return CompletableFuture.completedFuture(buildCityInfoContextResult(query, cityId));
-    }
-
-    /**
-     * De-duplicates sources by URL and preserves order: first occurrence wins,
-     * stable ordering.
-     */
-    public List<Source> deDuplicateAndOrderSources(List<Source> sources) {
-        LinkedHashSet<String> seenUrls = new LinkedHashSet<>();
-        List<Source> deduped = new ArrayList<>();
-        for (Source source : sources) {
-            if (seenUrls.add(source.getUrl())) {
-                deduped.add(source);
-            }
-        }
-        return List.copyOf(deduped);
-    }
-
-    private CityDetectionIndex getOrCreateDetectionIndex(String cityId) {
-        return detectionIndexByCity.computeIfAbsent(cityId, this::buildDetectionIndex);
-    }
-
-    private CityDetectionIndex buildDetectionIndex(String cityId) {
-        List<String> normalizedProcedureTitles = loadNormalizedProcedureTitles(cityId);
-        List<String> normalizedEventTitles = loadNormalizedEventTitles(cityId);
-        List<String> normalizedNewsTitles = loadNormalizedNewsTitles(cityId);
-        List<String> normalizedCityInfoTitles = loadNormalizedCityInfoTitles(cityId);
-
-        logger.info(() -> "Context detection index initialized — city=" + cityId
-                + " procedureTitles=" + normalizedProcedureTitles.size()
-                + " eventTitles=" + normalizedEventTitles.size()
-                + " newsTitles=" + normalizedNewsTitles.size()
-                + " cityInfoTitles=" + normalizedCityInfoTitles.size());
-
-        return new CityDetectionIndex(
-                TitleMatcher.fromTitles(normalizedProcedureTitles),
-                TitleMatcher.fromTitles(normalizedEventTitles),
-                TitleMatcher.fromTitles(normalizedNewsTitles),
-                TitleMatcher.fromTitles(normalizedCityInfoTitles),
-                normalizedProcedureTitles.size(),
-                normalizedEventTitles.size(),
-                normalizedNewsTitles.size(),
-                normalizedCityInfoTitles.size());
-    }
-
-    private List<String> loadNormalizedProcedureTitles(String cityId) {
-        try {
-            RagHelper<RagHelper.Procedure> helper = ragRegistry.getForCity(cityId);
-            return helper.getAll().stream()
-                    .map(procedure -> normalize(procedure.title))
-                    .filter(title -> !title.isBlank())
-                    .distinct()
-                    .toList();
-        } catch (Exception e) {
-            logger.fine(() -> "Failed to initialize procedure detection titles for city=" + cityId
-                    + " error=" + e.getMessage());
-            return List.of();
-        }
-    }
-
-    private List<String> loadNormalizedEventTitles(String cityId) {
-        try {
-            RagHelper<RagHelper.Event> helper = eventRagRegistry.getForCity(cityId);
-            return helper.getAll().stream()
-                    .map(event -> normalize(event.title))
-                    .filter(title -> !title.isBlank())
-                    .distinct()
-                    .toList();
-        } catch (Exception e) {
-            logger.fine(() -> "Failed to initialize event detection titles for city=" + cityId
-                    + " error=" + e.getMessage());
-            return List.of();
-        }
-    }
-
-    private List<String> loadNormalizedNewsTitles(String cityId) {
-        try {
-            RagHelper<RagHelper.News> helper = newsRagRegistry.getForCity(cityId);
-            return helper.getAll().stream()
-                    .map(item -> normalize(item.title))
-                    .filter(title -> !title.isBlank())
-                    .distinct()
-                    .toList();
-        } catch (Exception e) {
-            logger.fine(() -> "Failed to initialize news detection titles for city=" + cityId
-                    + " error=" + e.getMessage());
-            return List.of();
-        }
-    }
-
-    private List<String> loadNormalizedCityInfoTitles(String cityId) {
-        try {
-            RagHelper<RagHelper.CityInfo> helper = cityInfoRagRegistry.getForCity(cityId);
-            return helper.getAll().stream()
-                    .map(item -> normalize(item.title))
-                    .filter(title -> !title.isBlank())
-                    .distinct()
-                    .toList();
-        } catch (Exception e) {
-            logger.fine(() -> "Failed to initialize city-info detection titles for city=" + cityId
-                    + " error=" + e.getMessage());
-            return List.of();
-        }
-    }
-
-    private static String normalize(String value) {
-        if (value == null) {
-            return "";
-        }
-        return value.strip().toLowerCase(Locale.ROOT);
-    }
-
-    private static boolean isCandidateRelevantToQuery(RagHelper.Procedure procedure, String query) {
-        if (procedure == null || query == null || query.isBlank()) {
-            return false;
-        }
-
-        List<String> queryTokens = tokenizeForRelevance(query);
-        if (queryTokens.isEmpty()) {
-            return false;
-        }
-
-        Set<String> candidateTokens = new LinkedHashSet<>();
-        candidateTokens.addAll(tokenizeForRelevance(procedure.title));
-        candidateTokens.addAll(tokenizeForRelevance(procedure.description));
-
-        for (String token : queryTokens) {
-            if (candidateTokens.contains(token)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static List<String> tokenizeForRelevance(String text) {
-        if (text == null || text.isBlank()) {
-            return List.of();
-        }
-        return TokenNormalizer.tokenize(text).stream()
-                .filter(token -> token.length() >= 3)
-                .filter(token -> !TITLE_STOP_WORDS.contains(token))
-                .toList();
-    }
-
-    private static boolean containsAny(String normalizedQuestion, List<String> keywords) {
-        for (String keyword : keywords) {
-            if (normalizedQuestion.contains(keyword)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private static boolean isClearlyConversational(String normalizedQuestion) {
-        return containsAny(normalizedQuestion, CONVERSATIONAL_SHORT_CIRCUITS);
-    }
-
-    private static boolean hasDateWindow(String normalizedQuestion, String originalQuestion) {
-        if (containsAny(normalizedQuestion, DATE_WINDOW_KEYWORDS)) {
-            return true;
-        }
-
-        if (containsAny(normalizedQuestion, MONTH_KEYWORDS)) {
-            return true;
-        }
-
-        if (NUMERIC_DATE_PATTERN.matcher(originalQuestion).find()) {
-            return true;
-        }
-
-        if (ISO_DATE_PATTERN.matcher(originalQuestion).find()) {
-            return true;
-        }
-
-        return hasRangeExpressionWithDateTokens(normalizedQuestion, originalQuestion);
-    }
-
-    private static boolean hasRangeExpressionWithDateTokens(String normalizedQuestion, String originalQuestion) {
-        if (!containsAny(normalizedQuestion, RANGE_CONNECTOR_KEYWORDS)) {
-            return false;
-        }
-
-        return NUMERIC_DATE_PATTERN.matcher(originalQuestion).find()
-                || ISO_DATE_PATTERN.matcher(originalQuestion).find()
-                || containsAny(normalizedQuestion, MONTH_KEYWORDS)
-                || containsAny(normalizedQuestion, DATE_WINDOW_KEYWORDS);
-    }
-
-    private static Set<String> tokenize(String normalizedValue) {
-        if (normalizedValue == null || normalizedValue.isBlank()) {
-            return Set.of();
-        }
-
-        LinkedHashSet<String> tokens = new LinkedHashSet<>();
-        String[] rawTokens = normalizedValue.replaceAll("[^\\p{L}\\p{Nd}]+", " ").trim().split("\\s+");
-        for (String token : rawTokens) {
-            if (token.length() < 3 || TITLE_STOP_WORDS.contains(token)) {
-                continue;
-            }
-            tokens.add(token);
-        }
-        return tokens;
-    }
-}
