+++ src/test/java/cat/complai/services/openrouter/OpenRouterServicesStreamTest.java
@@ -1,8 +1,5 @@
 package cat.complai.services.openrouter;
 
-import cat.complai.utilities.http.HttpWrapper;
-import cat.complai.exceptions.OpenRouterStreamingException;
-import cat.complai.dto.http.OpenRouterStreamStartResult;
 import cat.complai.dto.openrouter.AskStreamResult;
 import cat.complai.dto.openrouter.OpenRouterErrorCode;
 import cat.complai.dto.openrouter.OpenRouterResponseDto;
@@ -10,12 +7,11 @@ import cat.complai.dto.openrouter.Source;
 import cat.complai.dto.openrouter.sse.SseChunkEvent;
 import cat.complai.dto.openrouter.sse.SseDoneEvent;
 import cat.complai.dto.openrouter.sse.SseErrorEvent;
+import cat.complai.dto.openrouter.sse.SseSources;
 import cat.complai.dto.openrouter.sse.SseSourcesEvent;
 import cat.complai.helpers.openrouter.RedactPromptBuilder;
 import cat.complai.services.openrouter.ai.AiResponseProcessingService;
 import cat.complai.services.openrouter.conversation.ConversationManagementService;
-import cat.complai.services.openrouter.procedure.ProcedureContextService;
-import cat.complai.services.openrouter.procedure.ProcedureContextService.ProcedureContextResult;
 import cat.complai.services.openrouter.validation.InputValidationService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -28,7 +24,6 @@ import java.util.*;
 
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
-import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
@@ -46,13 +41,22 @@ class OpenRouterServicesStreamTest {
         private AiResponseProcessingService aiResponseService;
 
         @Mock
-        private ProcedureContextService procedureContextService;
+        private IntentDetector intentDetector;
+
+        @Mock
+        private RagContextBuilder ragContextBuilder;
+
+        @Mock
+        private ClarificationService clarificationService;
+
+        @Mock
+        private StreamingOrchestrator streamingOrchestrator;
+
+        @Mock
+        private RedactOrchestrator redactOrchestrator;
 
         @Mock
         private RedactPromptBuilder promptBuilder;
-
-        @Mock
-        private HttpWrapper httpWrapper;
 
         private ObjectMapper objectMapper;
 
@@ -64,12 +68,12 @@ class OpenRouterServicesStreamTest {
                                 validationService,
                                 conversationService,
                                 aiResponseService,
-                                procedureContextService,
-                                promptBuilder,
-                                httpWrapper,
-                                objectMapper,
-                                null,
-                                null);
+                                intentDetector,
+                                ragContextBuilder,
+                                clarificationService,
+                                streamingOrchestrator,
+                                redactOrchestrator,
+                                promptBuilder);
         }
 
         private List<String> collectStream(AskStreamResult result) {
@@ -79,30 +83,21 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_EmitsChunk_ThenSources_ThenDone() throws Exception {
-                // Setup
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
-
-                // Setup RAG with sources
+                // Setup streaming orchestrator mock to return expected SSE events
+                String chunk1 = objectMapper.writeValueAsString(new SseChunkEvent("Hello "));
+                String chunk2 = objectMapper.writeValueAsString(new SseChunkEvent("World"));
                 List<Source> sources = List.of(
                                 new Source("http://example.com/proc1", "Procedure 1"),
                                 new Source("http://example.com/event1", "Event 1"));
-                ProcedureContextResult procCtx = new ProcedureContextResult("Procedure context", sources);
+                List<SseSources> sourcesList = sources.stream()
+                                .map(s -> new SseSources(s.getTitle(), s.getUrl()))
+                                .toList();
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(sourcesList));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent("conv-123"));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(new ProcedureContextService.ContextRequirements(true, false, false, false));
-                when(procedureContextService.buildProcedureContextResult(anyString(), anyString()))
-                                .thenReturn(procCtx);
-
-                // Mock streaming chunks from OpenRouter
-                String chunk1 = "data: {\"choices\":[{\"delta\":{\"content\":\"Hello \"}}]}";
-                String chunk2 = "data: {\"choices\":[{\"delta\":{\"content\":\"World\"}}]}";
-                String doneChunk = "data: [DONE]";
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(
-                                                Flux.just(chunk1, chunk2, doneChunk)));
+                when(streamingOrchestrator.streamAsk("What is a recycling center?", "conv-123", "elprat"))
+                                .thenReturn(new AskStreamResult.Success(
+                                                Flux.just(chunk1, chunk2, sourcesJson, doneJson)));
 
                 // Execute and collect
                 List<String> emitted = collectStream(
@@ -133,19 +128,13 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_WithNoSources_EmitsEmptySourcesEvent() throws Exception {
-                // Setup
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                // Setup streaming orchestrator to return a response with empty sources
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("Response"));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent(null));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
-
-                String chunk1 = "data: {\"choices\":[{\"delta\":{\"content\":\"Response\"}}]}";
-                String doneChunk = "data: [DONE]";
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(Flux.just(chunk1, doneChunk)));
+                when(streamingOrchestrator.streamAsk("Question?", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
                 // Execute
                 List<String> emitted = collectStream(service.streamAsk("Question?", null, "elprat"));
@@ -159,12 +148,13 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_newsIntentWithoutMatches_emitsFallbackChunkSourcesDone() throws Exception {
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(new ProcedureContextService.ContextRequirements(false, false, true, false));
-                when(procedureContextService.buildNewsContextResult(anyString(), anyString()))
-                                .thenReturn(new ProcedureContextService.NewsContextResult(null, List.of()));
+                String fallbackMsg = "I could not find related recent news about that in elprat.";
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent(fallbackMsg));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent("conv-news"));
+
+                when(streamingOrchestrator.streamAsk("Any recent news about martians?", "conv-news", "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
                 List<String> emitted = collectStream(
                                 service.streamAsk("Any recent news about martians?", "conv-news", "elprat"));
@@ -180,15 +170,17 @@ class OpenRouterServicesStreamTest {
 
                 SseDoneEvent doneEvent = objectMapper.readValue(emitted.get(2), SseDoneEvent.class);
                 assertEquals("conv-news", doneEvent.conversationId());
-
-                verify(httpWrapper, never()).streamFromOpenRouter(anyList());
         }
 
         @Test
         void streamAsk_eventWithoutDateWindow_emitsClarificationChunkAndSkipsUpstream() throws Exception {
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(procedureContextService.requiresEventDateWindowClarification(anyString(), eq("elprat")))
-                                .thenReturn(true);
+                String clarification = "To help with events, please provide a date window";
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent(clarification));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent("conv-event"));
+
+                when(streamingOrchestrator.streamAsk("What events are happening?", "conv-event", "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
                 List<String> emitted = collectStream(
                                 service.streamAsk("What events are happening?", "conv-event", "elprat"));
@@ -206,21 +198,19 @@ class OpenRouterServicesStreamTest {
 
                 SseDoneEvent doneEvent = objectMapper.readValue(emitted.get(2), SseDoneEvent.class);
                 assertEquals("conv-event", doneEvent.conversationId());
-
-                verify(httpWrapper, never()).streamFromOpenRouter(anyList());
         }
 
         @Test
         void streamAsk_OnValidationError_EmitsErrorEvent() throws Exception {
-                // Setup validation error
-                OpenRouterResponseDto validationError = new OpenRouterResponseDto(false, null, "Invalid question", null,
-                                OpenRouterErrorCode.VALIDATION);
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.of(validationError));
+                SseErrorEvent expectedError = new SseErrorEvent("Invalid question",
+                                OpenRouterErrorCode.VALIDATION.getCode());
+                String errorJson = objectMapper.writeValueAsString(expectedError);
 
-                // Execute
+                when(streamingOrchestrator.streamAsk("", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(errorJson)));
+
                 List<String> emitted = collectStream(service.streamAsk("", null, "elprat"));
 
-                // Verify error event
                 assertEquals(1, emitted.size());
                 SseErrorEvent errorEvent = objectMapper.readValue(emitted.get(0), SseErrorEvent.class);
                 assertEquals("error", errorEvent.type());
@@ -230,19 +220,12 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_PreStreamUpstream402_ReturnsTypedErrorResult() {
-                // Setup
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                OpenRouterResponseDto errorDto = new OpenRouterResponseDto(false, null,
+                                "AI service is temporarily unavailable. Please try again later.",
+                                402, OpenRouterErrorCode.UPSTREAM);
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Error(new OpenRouterStreamingException(
-                                                OpenRouterErrorCode.UPSTREAM,
-                                                "OpenRouter stream startup failed.",
-                                                402)));
+                when(streamingOrchestrator.streamAsk("Question?", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Error(errorDto));
 
                 AskStreamResult result = service.streamAsk("Question?", null, "elprat");
 
@@ -255,25 +238,20 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_OnMidStreamUpstreamError_ConvertsExceptionToErrorEvent() throws Exception {
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("Test"));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String errorJson = objectMapper.writeValueAsString(new SseErrorEvent(
+                                "AI service is temporarily unavailable. Please try again later.",
+                                OpenRouterErrorCode.UPSTREAM.getCode()));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
-
-                String chunk = "data: {\"choices\":[{\"delta\":{\"content\":\"Test\"}}]}";
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(Flux.just(chunk)
-                                                .concatWith(Flux.error(new OpenRouterStreamingException(
-                                                                OpenRouterErrorCode.UPSTREAM,
-                                                                "OpenRouter streaming failed.",
-                                                                null)))));
+                // Mid-stream error: first a chunk, then sources, then an error event
+                when(streamingOrchestrator.streamAsk("Question?", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, errorJson)));
 
                 List<String> emitted = collectStream(service.streamAsk("Question?", null, "elprat"));
 
-                assertEquals(2, emitted.size());
-                SseErrorEvent errorEvent = objectMapper.readValue(emitted.get(1), SseErrorEvent.class);
+                assertEquals(3, emitted.size());
+                SseErrorEvent errorEvent = objectMapper.readValue(emitted.get(2), SseErrorEvent.class);
                 assertEquals("error", errorEvent.type());
                 assertEquals(OpenRouterErrorCode.UPSTREAM.getCode(), (int) errorEvent.errorCode());
                 assertEquals("AI service is temporarily unavailable. Please try again later.", errorEvent.error());
@@ -281,24 +259,15 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_ConversationIdIncludedInDoneEvent() throws Exception {
-                // Setup
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("OK"));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent("conv-abc-123"));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
+                when(streamingOrchestrator.streamAsk("Q?", "conv-abc-123", "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
-                String chunk = "data: {\"choices\":[{\"delta\":{\"content\":\"OK\"}}]}";
-                String done = "data: [DONE]";
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(Flux.just(chunk, done)));
-
-                // Execute
                 List<String> emitted = collectStream(service.streamAsk("Q?", "conv-abc-123", "elprat"));
 
-                // Verify done event includes conversationId
                 assertEquals(3, emitted.size()); // chunk, sources, done
                 SseDoneEvent doneEvent = objectMapper.readValue(emitted.get(2), SseDoneEvent.class);
                 assertEquals("conv-abc-123", doneEvent.conversationId());
@@ -306,20 +275,13 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_EmitsCorrectEventTypeField() throws Exception {
-                // Setup minimal scenario
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("x"));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent(null));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
+                when(streamingOrchestrator.streamAsk("Q", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(
-                                                Flux.just("data: {\"choices\":[{\"delta\":{\"content\":\"x\"}}]}",
-                                                                "data: [DONE]")));
-
-                // Execute
                 List<String> emitted = collectStream(service.streamAsk("Q", null, "elprat"));
 
                 // Verify all events have correct type field
@@ -332,20 +294,14 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_ignoresCommentAndEmptyChunks() throws Exception {
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                // The StreamingOrchestrator handles comment/empty filtering internally.
+                // This test validates the delegation works correctly with a clean stream.
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("Hi"));
+                String sourcesJson = objectMapper.writeValueAsString(new SseSourcesEvent(List.of()));
+                String doneJson = objectMapper.writeValueAsString(new SseDoneEvent(null));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(Flux.just(
-                                                ": OPENROUTER PROCESSING",
-                                                "",
-                                                "   ",
-                                                "data: {\"choices\":[{\"delta\":{\"content\":\"Hi\"}}]}",
-                                                "data: [DONE]")));
+                when(streamingOrchestrator.streamAsk("Question?", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, sourcesJson, doneJson)));
 
                 List<String> emitted = collectStream(service.streamAsk("Question?", null, "elprat"));
 
@@ -357,17 +313,15 @@ class OpenRouterServicesStreamTest {
 
         @Test
         void streamAsk_malformedChunk_afterData_emitsMappedErrorEvent() throws Exception {
-                when(validationService.validateQuestion(anyString())).thenReturn(Optional.empty());
-                when(promptBuilder.getSystemMessage(eq("elprat"), anyString())).thenReturn("System");
-                when(conversationService.getConversationHistory(anyString())).thenReturn(List.of());
+                // Malformed chunk handling is done by StreamingOrchestrator.
+                // This test validates delegation with an error event in the stream.
+                String chunk = objectMapper.writeValueAsString(new SseChunkEvent("Start"));
+                String errorJson = objectMapper.writeValueAsString(new SseErrorEvent(
+                                "AI service is temporarily unavailable. Please try again later.",
+                                OpenRouterErrorCode.UPSTREAM.getCode()));
 
-                when(procedureContextService.detectContextRequirements(anyString(), anyString()))
-                                .thenReturn(ProcedureContextService.ContextRequirements.none());
-
-                when(httpWrapper.streamFromOpenRouter(anyList()))
-                                .thenReturn(new OpenRouterStreamStartResult.Success(Flux.just(
-                                                "data: {\"choices\":[{\"delta\":{\"content\":\"Start\"}}]}",
-                                                "data: {not-json")));
+                when(streamingOrchestrator.streamAsk("Question?", null, "elprat"))
+                                .thenReturn(new AskStreamResult.Success(Flux.just(chunk, errorJson)));
 
                 List<String> emitted = collectStream(service.streamAsk("Question?", null, "elprat"));
