+++ src/test/java/cat/complai/services/openrouter/OpenRouterServicesTest.java
@@ -10,7 +10,6 @@ import cat.complai.helpers.openrouter.ProcedureRagHelperRegistry;
 import cat.complai.helpers.openrouter.RagHelper;
 import cat.complai.services.openrouter.ai.AiResponseProcessingService;
 import cat.complai.services.openrouter.conversation.ConversationManagementService;
-import cat.complai.services.openrouter.procedure.ProcedureContextService;
 import cat.complai.services.openrouter.validation.InputValidationService;
 import org.junit.jupiter.api.Test;
 
@@ -22,6 +21,15 @@ import cat.complai.helpers.openrouter.RedactPromptBuilder;
 import cat.complai.dto.openrouter.OpenRouterErrorCode;
 import cat.complai.dto.openrouter.OutputFormat;
 import cat.complai.services.openrouter.cache.ResponseCacheService;
+import cat.complai.helpers.openrouter.CivicVocabularyService;
+import cat.complai.config.CivicVocabularyConfig;
+import cat.complai.services.openrouter.IntentDetector;
+import cat.complai.services.openrouter.RagContextBuilder;
+import cat.complai.services.openrouter.ClarificationService;
+import cat.complai.services.openrouter.StreamingOrchestrator;
+import cat.complai.services.openrouter.RedactOrchestrator;
+import cat.complai.helpers.openrouter.NewsRagHelperRegistry;
+import cat.complai.helpers.openrouter.CityInfoRagHelperRegistry;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -37,11 +45,22 @@ public class OpenRouterServicesTest {
         ConversationManagementService conversationService = new ConversationManagementService(5);
         ResponseCacheService cacheService = new ResponseCacheService(true, 10, 500);
         AiResponseProcessingService aiResponseService = new AiResponseProcessingService(wrapper, cacheService, 30);
-        ProcedureContextService procedureContextService = new ProcedureContextService(wrapper.ragRegistry,
-                new EventRagHelperRegistry(), new RedactPromptBuilder());
+        RedactPromptBuilder promptBuilder = new RedactPromptBuilder();
+        IntentDetector intentDetector = new IntentDetector(wrapper.ragRegistry, new EventRagHelperRegistry(),
+                new NewsRagHelperRegistry(), new CityInfoRagHelperRegistry());
+        RagContextBuilder ragContextBuilder = new RagContextBuilder(wrapper.ragRegistry, new EventRagHelperRegistry(),
+                new NewsRagHelperRegistry(), new CityInfoRagHelperRegistry(), promptBuilder);
+        ClarificationService clarificationService = new ClarificationService(wrapper.ragRegistry, intentDetector);
         com.fasterxml.jackson.databind.ObjectMapper objectMapper = new com.fasterxml.jackson.databind.ObjectMapper();
+        StreamingOrchestrator streamingOrchestrator = new StreamingOrchestrator(validationService,
+                conversationService, ragContextBuilder, clarificationService, intentDetector,
+                promptBuilder, wrapper, objectMapper);
+        RedactOrchestrator redactOrchestrator = new RedactOrchestrator(validationService,
+                conversationService, aiResponseService, promptBuilder,
+                new CivicVocabularyService(), new CivicVocabularyConfig());
         return new OpenRouterServices(validationService, conversationService, aiResponseService,
-                procedureContextService, new RedactPromptBuilder(), wrapper, objectMapper, null, null);
+                intentDetector, ragContextBuilder, clarificationService,
+                streamingOrchestrator, redactOrchestrator, promptBuilder);
     }
 
     // Flexible fake wrapper that simulates all test scenarios based on the last
@@ -245,11 +264,22 @@ public class OpenRouterServicesTest {
         ConversationManagementService conversationService = new ConversationManagementService(5);
         ResponseCacheService cacheService = new ResponseCacheService(true, 10, 500);
         AiResponseProcessingService aiResponseService = new AiResponseProcessingService(wrapper, cacheService, 30);
-        ProcedureContextService procedureContextService = new ProcedureContextService(wrapper.ragRegistry,
-                new EventRagHelperRegistry(), new RedactPromptBuilder());
+        RedactPromptBuilder promptBuilder = new RedactPromptBuilder();
+        IntentDetector intentDetector = new IntentDetector(wrapper.ragRegistry, new EventRagHelperRegistry(),
+                new NewsRagHelperRegistry(), new CityInfoRagHelperRegistry());
+        RagContextBuilder ragContextBuilder = new RagContextBuilder(wrapper.ragRegistry, new EventRagHelperRegistry(),
+                new NewsRagHelperRegistry(), new CityInfoRagHelperRegistry(), promptBuilder);
+        ClarificationService clarificationService = new ClarificationService(wrapper.ragRegistry, intentDetector);
         com.fasterxml.jackson.databind.ObjectMapper objectMapper = new com.fasterxml.jackson.databind.ObjectMapper();
+        StreamingOrchestrator streamingOrchestrator = new StreamingOrchestrator(validationService,
+                conversationService, ragContextBuilder, clarificationService, intentDetector,
+                promptBuilder, wrapper, objectMapper);
+        RedactOrchestrator redactOrchestrator = new RedactOrchestrator(validationService,
+                conversationService, aiResponseService, promptBuilder,
+                new CivicVocabularyService(), new CivicVocabularyConfig());
         OpenRouterServices svc = new OpenRouterServices(validationService, conversationService, aiResponseService,
-                procedureContextService, new RedactPromptBuilder(), wrapper, objectMapper, null, null);
+                intentDetector, ragContextBuilder, clarificationService,
+                streamingOrchestrator, redactOrchestrator, promptBuilder);
 
         OpenRouterResponseDto out = svc.redactComplaint("   ", OutputFormat.JSON, null, null, "testcity");
         assertFalse(out.isSuccess());
